% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adopters.R
\name{adopters}
\alias{adopters}
\title{Obtain Treatment from Design}
\usage{
adopters(design = NULL)
}
\arguments{
\item{design}{Optional design. If \code{NULL}, the Design will be identified
from its inclusion in a weight (\code{ate()} or \code{ett()}) or from a
covariate adjustment model (\code{cov_adj()}).}
}
\value{
The treatment variable to be placed in the regression formula.
}
\description{
Obtain Treatment from Design
}
