% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DirectAdjusted.R
\name{as.DirectAdjusted}
\alias{as.DirectAdjusted}
\title{Convert lm object into DirectAdjusted}
\usage{
as.DirectAdjusted(x, ...)
}
\arguments{
\item{x}{lm object with weights containing a WeightedDesign}

\item{...}{Add'l arguments}
}
\value{
DirectAdjusted
}
\description{
Convert lm object into DirectAdjusted
}
