% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SandwichLayer.R
\name{as.SandwichLayer}
\alias{as.SandwichLayer}
\title{Convert a PreSandwichLayer to a SandwichLayer via a Design Object}
\usage{
as.SandwichLayer(x, design, by = NULL, envir = parent.frame())
}
\arguments{
\item{x}{PreSandwichLayer}

\item{design}{Design}

\item{by}{character; must have entries corresponding to Design's
group assignment columns and names corresponding to the new desired column
names}
}
\value{
SandwichLayer
}
\description{
Convert a PreSandwichLayer to a SandwichLayer via a Design Object
}
