% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_center_residuals.R
\name{block_center_residuals}
\alias{block_center_residuals}
\title{Adjust residuals for both-sides absorption}
\usage{
block_center_residuals(x)
}
\arguments{
\item{x}{a fitted \code{teeMod} model}
}
\value{
the fitted \code{teeMod} with updated block center residuals.
}
\description{
Adjust residuals for both-sides absorption
}
\details{
This function subtracts off the block residual mean function
\eqn{\hat \alpha(v_b, \theta)} for each observation from model residuals
}
\keyword{internal}
