% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DirectAdjusted.R
\name{bread.DirectAdjusted}
\alias{bread.DirectAdjusted}
\title{Extract bread matrix from a \code{DirectAdjusted} model fit}
\usage{
\method{bread}{DirectAdjusted}(x)
}
\arguments{
\item{x}{a fitted \code{DirectAdjusted} object}
}
\value{
A \eqn{k\times k} matrix where k denotes the number of parameters
in the ITT effect model. This corresponds to the Hessian of the ITT effect
model estimating equations defined in our accompanying documentation (which
defines a perhaps unintuitive scaling factor of \eqn{|\{i : i \in Q\}| + |\{i : i in C\ Q|\}}
for this matrix).
}
\description{
Extract bread matrix from a \code{DirectAdjusted} model fit
}
