% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmrobMethods.R
\name{bread.glmrob}
\alias{bread.glmrob}
\title{Extract bread matrix from an \code{lmrob()} fit}
\usage{
\method{bread}{glmrob}(x, ...)
}
\arguments{
\item{x}{a fitted \code{lmrob} object}

\item{...}{arguments passed to methods}
}
\value{
Details needed
}
\description{
Extract bread matrix from an \code{lmrob()} fit
}
