% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SandwichLayerVariance.R
\name{bread.lmrob}
\alias{bread.lmrob}
\title{Bread method for lmrob objects}
\usage{
\method{bread}{lmrob}(x, ...)
}
\arguments{
\item{x}{An \code{lmrob} object produced using an MM/SM estimator chain}

\item{...}{Additional arguments to be passed to \code{bread}}
}
\value{
A \eqn{p\times }(p+1) matrix where the first column corresponds to
the scale estimate and the remaining \eqn{p} colums correspond to the
coefficients
}
\description{
Bread method for lmrob objects
}
\details{
Extract bread matrix from an \code{lmrob} fit

This is part of a workaround for an issue in the robustbase code
affecting sandwich covariance estimation.
The issue in question is issue #6471, robustbase project on R-Forge.
}
\author{
lrd author 2
}
