% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teeMod.R
\name{bread.teeMod}
\alias{bread.teeMod}
\title{Extract bread matrix from a \code{teeMod} model fit}
\usage{
\method{bread}{teeMod}(x, ...)
}
\arguments{
\item{x}{a fitted \code{teeMod} model}

\item{...}{arguments passed to methods}
}
\value{
A \eqn{2\times 2} matrix corresponding to an intercept and the
treatment variable in the direct adjustment model
}
\description{
An S3method for \code{sandwich::bread} that extracts the bread of the
direct adjustment model sandwich covariance matrix.
}
\details{
This function is a thin wrapper around \code{.get_tilde_a22_inverse()}.
}
