% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CombinedWeightedDesign.R
\name{c,WeightedDesign-method}
\alias{c,WeightedDesign-method}
\title{Concatenate weights}
\usage{
\S4method{c}{WeightedDesign}(x, ..., force_dichotomy_equal = FALSE)
}
\arguments{
\item{x, }{.. a \code{WeightedDesign} object, typically created from \code{\link[=ate]{ate()}}
or \code{\link[=ett]{ett()}}}

\item{...}{any number of additional \code{WeightedDesign} objects with
equivalent \code{Design} to \code{x} and eachother}

\item{force_dichotomy_equal}{if \code{FALSE} (default), \code{WeightedDesign}s are
considered equivalent even if their \code{dichotomy} differs. If
\code{TRUE}, \code{@dichotomy} must also be equal.}
}
\value{
A new \code{WeightedDesign} with
the weights concatenated in the input order.
}
\description{
Given several variations of weights generated from a single
\code{Design}, combine into a single weight.
}
\details{
Concatenating \code{WeightedDesign} objects with \code{\link[=c]{c()}} requires
both individual \code{WeightedDesign} objects to come from the same
\code{Design} (except \code{dichotomy}, see below) and have the same
target (e.g all created with \code{\link[=ate]{ate()}} or all created with \code{\link[=ett]{ett()}}, no
mixing-and-matching). All arguments to \code{\link[=c]{c()}} must be
\code{WeightedDesign}.

One exception is when concatenting \code{WeightedDesign} objects whose
\code{Design} differ only in their dichotomies. There may be cases where
the treatment is continuous or has multiple levels, and there is a need to
combine the weights from the same general design, but with different
dichotomies. Therefore multiple \code{WeightedDesign} objects can be
combined if they are identical except for their \code{@dichotomy} slots.
The resulting object will be a \code{WeightedDesign}.
}
\examples{
data(simdata)
des <- rct_design(z ~ unit_of_assignment(uoa1, uoa2), data = simdata)
w1 <- ate(des, data = simdata[1:30,])
w2 <- ate(des, data = simdata[31:40,])
w3 <- ate(des, data = simdata[41:50,])
c_w <- c(w1, w2, w3)
c(length(w1), length(w2), length(w3), length(c_w))

des <- rct_design(dose ~ unit_of_assignment(uoa1, uoa2), data = simdata)
w1 <- ate(des, data = simdata[1:10, ], dichotomy = dose >= 300 ~ .)
w2 <- ate(des, data = simdata[11:30, ], dichotomy = dose >= 200 ~ .)
w3 <- ate(des, data = simdata[31:50, ], dichotomy = dose >= 100 ~ .)
c_w <- c(w1, w2, w3)
}
