% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DirectAdjusted.R
\name{confint,DirectAdjusted-method}
\alias{confint,DirectAdjusted-method}
\title{Variance-Covariance matrix}
\usage{
\S4method{confint}{DirectAdjusted}(object, parm, level = 0.95, ...)
}
\arguments{
\item{object}{DirectAdjusted}

\item{...}{Add'l arguments}
}
\value{
Variance-Covariance matrix
}
\description{
Variance-Covariance matrix
}
