% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lmitted.R
\name{confint,Lmitted-method}
\alias{confint,Lmitted-method}
\title{Variance-Covariance matrix}
\usage{
\S4method{confint}{Lmitted}(object, parm, level = 0.95, ...)
}
\arguments{
\item{object}{Lmitted}

\item{parm}{a specification of which parameters are to be given confidence
intervals, either a vector of numbers or a vector of names. If missing,
all parameters are considered.}

\item{level}{the confidence level required.}

\item{...}{Add'l arguments}
}
\value{
Variance-Covariance matrix
}
\description{
Variance-Covariance matrix
}
