% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DirectAdjusted.R
\name{confint.DirectAdjusted}
\alias{confint.DirectAdjusted}
\title{Variance-Covariance matrix}
\usage{
\method{confint}{DirectAdjusted}(object, parm, type = c("CR1"), level = 0.95, ...)
}
\arguments{
\item{object}{DirectAdjusted}

\item{parm}{a specification of which parameters are to be given confidence
intervals, either a vector of numbers or a vector of names. If missing,
all parameters are considered.}

\item{type}{A string indicating the desired variance estimator. Currently
accepts "CR1".}

\item{level}{the confidence level required.}

\item{...}{Add'l arguments}
}
\value{
Variance-Covariance matrix
}
\description{
Variance-Covariance matrix
}
