% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_adj.R
\name{default}
\alias{default}
\title{Instruct \code{cov_adj()} to find a default reference value for columns in
a covariance adjustment model}
\usage{
default()
}
\value{
empty list that inherits from class \code{set_to_reference_default}
}
\description{
Instruct \code{cov_adj()} to find a default reference value for columns in
a covariance adjustment model
}
\details{
When set as the entry of a list passed to the \code{set_to_reference}
argument of \code{cov_adj()}, \code{cov_adj()} uses the following
replacement values depending on the column type:
\itemize{
\item factor/character: minimum level
\item numeric: minimum value
\item logical: FALSE
}
}
