% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DesignAccessors.R
\name{dichotomization}
\alias{dichotomization}
\alias{dichotomization,Design-method}
\alias{dichotomization<-}
\alias{dichotomization<-,Design-method}
\title{Extract or replace dichotomization}
\usage{
dichotomization(x)

\S4method{dichotomization}{Design}(x)

dichotomization(x) <- value

\S4method{dichotomization}{Design}(x) <- value
}
\arguments{
\item{x}{Design object}
}
\value{
Dichomization formula
}
\description{
Extract or replace dichotomization
}
