% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DesignAccessors.R
\name{dichotomy}
\alias{dichotomy}
\alias{dichotomy,Design-method}
\alias{dichotomy<-}
\alias{dichotomy<-,Design-method}
\title{Extract or replace dichotomy}
\usage{
dichotomy(x)

\S4method{dichotomy}{Design}(x)

dichotomy(x) <- value

\S4method{dichotomy}{Design}(x) <- value
}
\arguments{
\item{x}{a \code{Design} object}

\item{value}{replacement \code{dichotomy} formula, or \code{NULL} to remove}
}
\value{
dichomization formula, or a \code{Design} with the
\code{dichomization} replaces with \code{x}.
}
\description{
Extract or replace dichotomy
}
