% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DesignStructure.R
\name{get_structure}
\alias{get_structure}
\alias{show,DesignStructure-method}
\title{\code{Design} Structure Information}
\usage{
get_structure(design)

\S4method{show}{DesignStructure}(object)
}
\arguments{
\item{design}{a \code{Design} object}

\item{object}{a \code{DesignStructure} object, typically the output of
\code{\link[=get_structure]{get_structure()}}}
}
\value{
A \code{DesignStructure} object containing the structure of the
\code{design} as a \code{data.frame}.
}
\description{
Obtaining a \code{data.frame} which encodes the design
information.
}
\examples{
data(simdata)
des <- rct_design(z ~ uoa(uoa1, uoa2) + block(bid), data = simdata)
get_structure(des)
}
