% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.Design.R
\name{summary.Design}
\alias{summary.Design}
\alias{print.summary.Design}
\title{Summarizing \code{Design} objects}
\usage{
\method{summary}{Design}(object, ..., treatment_binary = TRUE)

\method{print}{summary.Design}(x, ..., max_unit_print = 3)
}
\arguments{
\item{object}{\code{Design} object, usually a result of a call to
\code{\link[=rct_design]{rct_design()}}, \code{\link[=obs_design]{obs_design()}}, or \code{\link[=rd_design]{rd_design()}}.}

\item{...}{Ignored}

\item{treatment_binary}{Should the treatment be dichotomized if
\code{object} contains a \code{dichotomy}? Ignored if \code{object} does
not contain a \code{dichotomy}.}

\item{x}{\code{summary.Design} object, usually as a result of a call to
\code{\link[=summary.Design]{summary.Design()}}}

\item{max_unit_print}{Maximum number of treatment levels to print in
treatment table}
}
\value{
The \code{Design} or \code{summary.Design}object, invisibly
}
\description{
\code{\link[=summary]{summary()}} method for class \code{Design}.
}
