% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design.R
\name{var_table}
\alias{var_table}
\alias{var_names}
\title{Extract Variable Names from \code{Design}}
\usage{
var_table(design, compress = TRUE, report_all = FALSE)

var_names(design, type)
}
\arguments{
\item{design}{a \code{Design} object}

\item{compress}{should multiple variables be compressed into a
comma-separated string? Default \code{TRUE}. If \code{FALSE}, multiple
columns can be created instead.}

\item{report_all}{should we report all possible structures even if they
don't exist in the \code{Design}? Default \code{FALSE}.}

\item{type}{one of "t", "u", "b", "f"; for "treatment",
"unit_of_assignment", "block", and "forcing" respectively}
}
\value{
\code{\link[=var_table]{var_table()}} returns the requested table. \code{\link[=var_names]{var_names()}} returns a
vector of variable names.
}
\description{
Methods to extract the variable names to the elements of the
structure of the \code{Design} (e.g. treatment, unit of analysis, etc)
}
\details{
When \code{compress} is \code{TRUE}, the result will always have
two columns. When \code{FALSE}, the result will have number of columns
equal to the largest number of variables in a particular role, plus one.
E.g., a call such as \code{rct_design(z ~ unitid(a, b, c, d) ...} will
have 4+1=5 columns in the output matrix with \code{compress = FALSE}.

When \code{report_all} is \code{TRUE}, the matrix is guaranteed to have 3
rows (when the \code{design} is an RCT or Obs) or 4 rows (when the
\code{design} is a RD), with empty variable entries as appropriate. When
\code{FALSE}, the matrix will have minimum 2 rows (treatment and unit of
assignment/unitid/cluster), with additional rows for blocks and forcing if
included in the \code{Design}.
}
\examples{
des <- rct_design(z ~ uoa(cid1, cid2) + block(bid), data = simdata)
var_table(des)
var_table(des, compress = FALSE)
var_names(des, "t")
var_names(des, "u")
var_names(des, "b")
}
