% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design.R
\name{forcing}
\alias{forcing}
\alias{cluster}
\title{Special terms in Design}
\usage{
forcing(...)

cluster(...)

forcing(...)
}
\arguments{
\item{...}{any number of variables of the same length.}
}
\value{
the variables with appropriate labels
}
\description{
These are special function used only in the definition of Design class
objects. They identify the clusters, blocks and forcing variables.
}
