% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dichotomize_design.R
\name{dichotomize_design}
\alias{dichotomize_design}
\title{Add dichotomization to a design}
\usage{
dichotomize_design(des, dichotomize, force = TRUE)
}
\arguments{
\item{des}{a Design object}

\item{dichotomize}{optionally, a formula defining the dichotomization of the
treatment variable if it isn't already \code{0}/\code{1}. See details of
help for \code{\link{rct_design()}} e.g. for details.}

\item{force}{If \code{des} is already dichotomized, should we force a new
dichomization (TRUE, default) or should we instead return the original
design (FALSE)?}
}
\value{
A dichotomized design
}
\description{
Generally users should leave \code{force = TRUE}. However, if you want to avoid
accidentally changing the dichotomization, set \code{force = FALSE}.
}
