% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.DirectAdjusted.R
\name{summary.DirectAdjusted}
\alias{summary.DirectAdjusted}
\alias{print.summary.DirectAdjusted}
\title{Summarizing \code{DirectAdjusted} objects}
\usage{
\method{summary}{DirectAdjusted}(object, vcov.type = "CR0", ...)

\method{print}{summary.DirectAdjusted}(
  x,
  digits = max(3L, getOption("digits") - 3L),
  signif.stars = getOption("show.signif.stars"),
  ...
)
}
\arguments{
\item{object}{\code{DirectAdjusted} object}

\item{vcov.type}{A string indicating the desired variance estimator. See
\code{\link[=vcovDA]{vcovDA()}} for details on accepted types.}

\item{...}{Additional arguments to \code{\link[=vcovDA]{vcovDA()}}, such as the desired
finite sample heteroskedasticity-robust standard error adjustment.}

\item{x}{\code{summary.DirectAdjusted} object}

\item{digits}{the number of significant digits to use when printing.}

\item{signif.stars}{logical. If ‘TRUE’, ‘significance stars’ are printed for
each coefficient.}
}
\value{
object of class \code{summary.DirectAdjusted}
}
\description{
\code{\link[=summary]{summary()}} method for class \code{DirectAdjusted}
}
\details{
If a \code{DirectAdjusted} object is fit with a
\code{SandwichLayer} offset, then the usual \code{stats::summary.lm()}
output is enhanced by the use of covariance-adjusted sandwich standard
errors, with t-test values recalculated to reflect the new standard
errors.
}
