% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StudySpecificationAccessors.R
\name{..uoa..}
\alias{..uoa..}
\title{Return ..uoa.. column}
\usage{
..uoa..(spec)
}
\arguments{
\item{spec}{A studyspecification}
}
\value{
The ..uoa.. column
}
\description{
Return ..uoa.. column
}
\keyword{internal}
