% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmitt.R
\name{.absorbed}
\alias{.absorbed}
\title{(Internal) Include blocks as absorbed effects}
\usage{
.absorbed(x)
}
\arguments{
\item{x}{Block variables}
}
\value{
Identifies block variables as categorical for the model.
}
\description{
Not to be used interactively. Alias for \code{as.factor()} to identify fixed
effects generated through the \code{absorb=TRUE} option to \code{lmitt()}.
}
