% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SandwichLayerVariance.R
\name{.add_mat_sqdif}
\alias{.add_mat_sqdif}
\title{(Internal) Helper function for design-based meat matrix
calculation}
\usage{
.add_mat_sqdif(X, zobs, bid, b, upper = TRUE)
}
\description{
(Internal) Helper function for design-based meat matrix
calculation
}
\keyword{internal}
