% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SandwichLayerVariance.R
\name{.add_vec}
\alias{.add_vec}
\title{(Internal) Helper function for design-based meat matrix calculation}
\usage{
.add_vec(a, upper = TRUE)
}
\description{
(Internal) Helper function for design-based meat matrix calculation
}
\keyword{internal}
