% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SandwichLayerVariance.R
\name{.aggregate_to_cluster}
\alias{.aggregate_to_cluster}
\title{(Internal) Aggregate weights and outcomes to cluster level}
\usage{
.aggregate_to_cluster(x, ...)
}
\arguments{
\item{x}{a fitted \code{teeMod} model}
}
\value{
a list of
\itemize{
\item a data frame of cluster weights, outcomes, treatments, and block ids;
\item treatment id column name;
\item block id column name
}
}
\description{
(Internal) Aggregate weights and outcomes to cluster level
}
\details{
aggregate individual weights and outcomes to cluster weighted sums
}
\keyword{internal}
