% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DesignAccessors.R
\name{.apply_dichotomy}
\alias{.apply_dichotomy}
\title{(Internal) Applies dichotomy to treatment}
\usage{
.apply_dichotomy(txt, dichotomy)
}
\arguments{
\item{txt}{A named \code{data.frame} containing a single column of the
treatment, such as that produed by \code{treatment(mydesign)}.}

\item{dichotomy}{A dichotomization formula. See the details in the Details
for the help of \code{rct_design()}.}
}
\value{
A \code{vector} of binary treatments
}
\description{
Given a treatment variable (passed as a named \code{data.frame}) and a
dichotomy formula (see help on \code{rct_design()} for details on
specification), returns \code{vector} containing only \code{0}, \code{1}, or
\code{NA}.
}
\keyword{internal}
