% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_data_frame.R
\name{.as_data_frame}
\alias{.as_data_frame}
\title{Convert object to \code{data.frame} or produce meaningful error}
\usage{
.as_data_frame(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{x} as a \code{data.frame}
}
\description{
Convert object to \code{data.frame} or produce meaningful error
}
\keyword{internal}
