% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DirectAdjusted.R
\name{.base_S3class_estfun}
\alias{.base_S3class_estfun}
\title{(Internal) Call \code{sandwich::estfun} method for a fitted \code{DirectAdjusted}
object's based on its base S3 class}
\usage{
.base_S3class_estfun(x)
}
\arguments{
\item{x}{a fitted \code{DirectAdjusted} object}
}
\value{
S3 method
}
\description{
(Internal) Call \code{sandwich::estfun} method for a fitted \code{DirectAdjusted}
object's based on its base S3 class
}
\keyword{internal}
