% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DesignAccessors.R
\name{.bin_txt}
\alias{.bin_txt}
\title{(Internal) Extracts treatment as binary \code{vector} if possible or
else errors.}
\usage{
.bin_txt(des)
}
\arguments{
\item{des}{A \code{Design}}
}
\value{
A \code{vector} of binary treatments
}
\description{
If the \code{Design} has a \code{@dichotomy}, or has a treatment variable
consisting only of 0/1 or \code{NA}, then returns the binary treatment.
Otherwise (it has a non-binary treatment and lacks a dichotomy) it errors.
}
\keyword{internal}
