% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DesignAccessors.R
\name{.binarize_treatment}
\alias{.binarize_treatment}
\title{(Internal) Uses a \code{@dichotomy} to create a binary version of the}
\usage{
.binarize_treatment(trt, dichot)
}
\arguments{
\item{trt}{A named \code{data.frame} containing a single column of the
treatment, such as that produed by \code{treatment(mydesign)}.}

\item{dichot}{A dichotomization formula. See the details in the Details for
the help of \code{rct_design()}.}
}
\value{
A \code{vector} of binary treatments
}
\description{
(Internal) Uses a \code{@dichotomy} to create a binary version of the
}
\keyword{internal}
