% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DirectAdjusted.R
\name{.calc_fe_correction}
\alias{.calc_fe_correction}
\title{Calculate meat matrix correction under fixed effects absorption}
\usage{
.calc_fe_correction(x, cluster = NULL, by = NULL, ...)
}
\arguments{
\item{x}{a fitted \code{DirectAdjusted} object}

\item{cluster}{character vector or list; optional. Specifies column names that appear
in both the covariance adjustment dataframe C and quasiexperimental dataframe
Q. Defaults to NULL, in which case unit of assignment columns indicated in
the Design will be used to generate clustered covariance estimates.}

\item{by}{character vector or list; optional. Specifies column names that appear in
botn the covariance adjustment dataframe C and quasiexperimental dataframe Q.
Defaults to NULL, in which case unit of assignment columns indicated in the
Design will be used to generate ID's.}

\item{...}{arguments passed to methods}
}
\description{
Calculate meat matrix correction under fixed effects absorption
}
