% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DirectAdjusted.R
\name{.calc_hc_corrections}
\alias{.calc_hc_corrections}
\title{Calculate HC corrections}
\usage{
.calc_hc_corrections(psi, phi = NULL, correction_type = NULL, ...)
}
\arguments{
\item{psi}{matrix, estimating equations for the ITT effect model}

\item{phi}{matrix, estimating equations for the covariance adjustment model
model}

\item{correction_type}{character, HC correction type for the variance
estimate. Default is NULL, which returns no correction (an HC0 estimate)}
}
\description{
Calculate HC corrections
}
