% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_by.R
\name{.check_by}
\alias{.check_by}
\title{(Internal) A few checks to ensure \code{by=} is valid}
\usage{
.check_by(by)
}
\arguments{
\item{by}{named vector or list connecting names of unit of
assignment/unitid/cluster variables in \code{specification} to unit of
assignment/unitid/cluster variables in \code{data}. Names represent
variables in the StudySpecification; values represent variables in the
data.}
}
\value{
\code{NULL} if no errors are found
}
\description{
Thie ensures that the \code{by=} argument is of the proper type, is named,
and consists of only unique entries.
}
\keyword{internal}
