% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SandwichLayerVariance.R
\name{.check_cluster_col_nas}
\alias{.check_cluster_col_nas}
\title{(Internal) Check the NA status of clustering columns in a dataframe}
\usage{
.check_cluster_col_nas(
  cluster_cols,
  df,
  model = c("covariance adjustment", "ITT effect")
)
}
\arguments{
\item{cluster_cols}{vector of column names}

\item{df}{dataframe whose columns are checked for NA's}

\item{model_type}{string taking either the value "covariance adjustment",
referring to the covariance adjustment model, or "ITT effect", referring to
the direct adjusted model. This value is piped into any potential warning
message.}
}
\value{
A vector of column names for the valid clustering columns
}
\description{
(Internal) Check the NA status of clustering columns in a dataframe
}
\details{
\code{.check_cluster_col_nas} checks whether entire columns in a
dataframe contain NA values. If a column only contains NA's, the function
produces a warning that the given indicating the column cannot be used for
clustering.
}
\keyword{internal}
