% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DesignUtilities.R
\name{.check_design_formula}
\alias{.check_design_formula}
\title{(Internal) Perform checks on formula for creation of Design.}
\usage{
.check_design_formula(form, allow_forcing = FALSE)
}
\arguments{
\item{form}{A formula passed to \code{.new_Design()}}

\item{allow_forcing}{Binary whether \code{forcing()} is allowed (\code{TRUE}
for RDD, \code{FALSE} for RCT and Obs).}
}
\value{
\code{TRUE} if all checks pass, otherwise errors.
}
\description{
Checks performed:
\itemize{
\item Ensure presence of exactly one of \code{unit_of_assignment()},
\code{cluster()} or \code{unitid()}.
\item Disallow multiple \code{block()} or multiple \code{forcing()} terms.
\item Disallow \code{forcing()} unless in RDD.
}
}
\keyword{internal}
