% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SandwichLayerVariance.R
\name{.check_df_moderator_estimates}
\alias{.check_df_moderator_estimates}
\title{(Internal) Replace standard errors for moderator effect estimates with insufficient
degrees of freedom with \code{NA}}
\usage{
.check_df_moderator_estimates(
  vmat,
  model,
  cluster,
  model_data = quote(data),
  envir = environment(formula(model))
)
}
\arguments{
\item{vmat}{output of \code{.vcov_XXX()} called with input to \code{model}
argument below as the first argument}

\item{model}{a fitted \code{teeMod} model}

\item{cluster}{a character or factor vector denoting clusters the units of
observation used to fit \code{model} belong to. Most conveniently, the
output of \code{.make_uoa_ids()}}

\item{model_data}{dataframe or name of dataframe used to fit \code{model}}

\item{envir}{environment to get \code{model_data} from if \code{model_data} has class \code{name}}
}
\value{
A variance-covariance matrix with NA's for estimates lacking sufficient
degrees of freedom
}
\description{
(Internal) Replace standard errors for moderator effect estimates with insufficient
degrees of freedom with \code{NA}
}
\keyword{internal}
