% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teeMod.R
\name{.compute_loo_resids}
\alias{.compute_loo_resids}
\title{Compute residuals for a \code{teeMod} object with leave-one-out estimates
of the \code{offset}}
\usage{
.compute_loo_resids(x, cluster, ...)
}
\arguments{
\item{x}{a \code{teeMod} object}

\item{cluster}{vector of column names that identify clusters}
}
\description{
Compute residuals for a \code{teeMod} object with leave-one-out estimates
of the \code{offset}
}
\details{
The residual for any observation also used for fitting the \code{fitted_covariance_model}
stored in the \code{offset} of \code{x} is replaced by an estimated residual
that uses a cluster leave-one-out estimate of the \code{fitted_covariance_model}
for generating a value of the \code{offset}.
}
\keyword{internal}
