% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DesignAccessors.R
\name{.convert_to_data.frame}
\alias{.convert_to_data.frame}
\title{(Internal) Ensures replacement column for \code{Design} is a
\code{data.frame}.}
\usage{
.convert_to_data.frame(value, design, type)
}
\arguments{
\item{value}{A \code{vector} or \code{data.frame} containing a replacement.}

\item{design}{A \code{Design}}

\item{type}{One of "t", "f", "u" or "b"}
}
\value{
\code{data.frame} containing named column(s)
}
\description{
Helper function for \code{Design} replacers.
}
\details{
When given a replacement set of values (either a \code{vector} or a
\code{data.frame}), this ensures that the replacement is a
\code{data.frame}.

Additionally ensures proper names for replacement
}
\keyword{internal}
