% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SandwichLayerVariance.R
\name{.cov01_est}
\alias{.cov01_est}
\title{(Internal) Helper function for design-based meat matrix calculation}
\usage{
.cov01_est(XX, zobs, bid)
}
\description{
(Internal) Helper function for design-based meat matrix calculation
}
\keyword{internal}
