% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SandwichLayerVariance.R
\name{.cov_mat_est}
\alias{.cov_mat_est}
\title{(Internal) Helper function for design-based meat matrix calculation}
\usage{
.cov_mat_est(XXz, bidz)
}
\description{
(Internal) Helper function for design-based meat matrix calculation
}
\keyword{internal}
