% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DesignAccessors.R
\name{.design_accessors_newdata_validate}
\alias{.design_accessors_newdata_validate}
\title{(Internal) Checks newdata/by argument for design accessors}
\usage{
.design_accessors_newdata_validate(newdata, by)
}
\arguments{
\item{newdata}{newdata argument from e.g. \code{treatment()},
\code{blocks()}, etc}

\item{by}{from e.g. \code{treatment()},
\code{blocks()}, etc. See \code{.check_by()}}
}
\value{
Invisibly \code{TRUE}. Warns or errors as appropriate.
}
\description{
(Internal) Checks newdata/by argument for design accessors
}
\keyword{internal}
