% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teeMod.R
\name{.estfun_DB_blockabsorb}
\alias{.estfun_DB_blockabsorb}
\title{Design-based estimating equations contributions}
\usage{
.estfun_DB_blockabsorb(x, by = NULL, ...)
}
\arguments{
\item{x}{a fitted \code{teeMod} object}

\item{...}{arguments passed to methods}
}
\value{
An \eqn{n\times k} matrix
}
\description{
Design-based estimating equations contributions
}
\details{
calculate contributions to empirical estimating equations from a
\code{teeMod} model with absorbed intercepts from the design-based
perspective
}
\keyword{internal}
