% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teeMod.R
\name{.estfun_psi_absorbed}
\alias{.estfun_psi_absorbed}
\title{(Internal) Absorb block effects into effect estimating equations by
estimating them with a random effects model}
\usage{
.estfun_psi_absorbed(x, weights)
}
\arguments{
\item{x}{a fitted \code{teeMod} model}

\item{weights}{numeric vector of the weights in the model of the covariance
structure. If NULL, uniform weights will be used}
}
\value{
a matrix with number of rows given by the number of rows of
\code{stats::model.frame(x)} and number of columns given by the number of
parameters in the model
}
\description{
(Internal) Absorb block effects into effect estimating equations by
estimating them with a random effects model
}
\details{
Block effects are estimated by modeling the covariance of the errors
using a random effects structure. Errors within blocks specified by the
\code{Design} slot of the \code{teeMod} are modeled as being correlated, while
errors across blocks are modeled as being uncorrelated. Residual variances of
the errors are modeled as heteroskedastic but proportional to weights provided
to the \code{weights} argument. The elements of the covariance matrix are
estimated following the approach in Imbens and Kolesar (2016).
}
\references{
Imbens, Guido W., and Michal Kolesar. "Robust standard
errors in small samples: Some practical advice." \emph{Review of Economics
and Statistics} 98.4 (2016): 701-712.
}
\keyword{internal}
