% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StudySpecificationAccessors.R
\name{.expand_txt}
\alias{.expand_txt}
\title{(Internal) Expand treatment variable from a \code{StudySpecification}
to a dataframe with unit of assignment information}
\usage{
.expand_txt(txt, data, spec)
}
\arguments{
\item{txt}{A dataframe with one column corresponding to the treatment. Can
be dichotomized or as it's stored in \code{spec}}

\item{data}{A dataframe with unit of assignment information}

\item{spec}{A \code{StudySpecification}, used to align unit of assignment
information with \code{txt}}
}
\description{
(Internal) Expand treatment variable from a \code{StudySpecification}
to a dataframe with unit of assignment information
}
\keyword{internal}
