% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_from_model.R
\name{.fallback_data_search}
\alias{.fallback_data_search}
\title{(Internal) Fallback brute force method to locate \code{data} in the
call stack.}
\usage{
.fallback_data_search()
}
\value{
If found, the data.
}
\description{
We try to be intelligent about finding the appropriate data. If this fails,
we may have need for a brute force method that just loops through frames and
looks for a \code{data} object.
}
\keyword{internal}
