% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dichotomy.R
\name{.find_dichotomies}
\alias{.find_dichotomies}
\title{(Internal) Find \code{dichotomy} formulas in the call stack}
\usage{
.find_dichotomies()
}
\value{
A list where elements are either formulas or NULL, depending on
whether a \code{dichotomy} argument was found in \code{lmitt.formula()} or
its \code{weights} argument
}
\description{
(Internal) Find \code{dichotomy} formulas in the call stack
}
\details{
\code{.find_dichotomies()} searches for \code{lmitt.formula()}
calls and their \code{weights} arguments for any \code{dichotomy}
arguments.
}
\keyword{internal}
