% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SandwichLayerVariance.R
\name{.get_a11_inverse}
\alias{.get_a11_inverse}
\title{(Internal) Get the inverse of the A11 block of the sandwich variance estimator}
\usage{
.get_a11_inverse(x)
}
\arguments{
\item{x}{A \code{Lmitted} model}
}
\value{
a (p+1)x(p+1) matrix where the dimensions are given by the number of
terms in the covariance model (p) and an Intercept term.
}
\description{
(Internal) Get the inverse of the A11 block of the sandwich variance estimator
}
\details{
This block is the pxp matrix corresponding to the unscaled
inverse of the observed Fisher information of the covariance model. The
observed information is given by the estimate of the negative Jacobian of the
model's estimating equations. The unscaled version provided here divides by
the number of observations used to fit the covariance model.
}
\keyword{internal}
