% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SandwichLayerVariance.R
\name{.get_a22_inverse}
\alias{.get_a22_inverse}
\title{(Internal) Get the inverse of the A22 block of the sandwich variance estimator}
\usage{
.get_a22_inverse(x)
}
\arguments{
\item{x}{A \code{Lmitted} model}
}
\value{
A 1x1 matrix
}
\description{
(Internal) Get the inverse of the A22 block of the sandwich variance estimator
}
\details{
This block is the diagonal element of the inverse expected Fisher
Information matrix corresponding to the treatment estimate. As shown in the
Details of \code{\link{.get_b22}}, the estimating equations for a generalized
linear model with a canonical link function can be written as
\deqn{\psii = (ri / Var(yi)) * (d\mui/d\etai) * xi}
The expected information matrix A22 is then the negative Jacobian of \eqn{\psii},
which by likelihood theory is the variance-covariance matrix of \eqn{\psii}:
\deqn{\psi_{i}\psi_{i}^{T} = E[(ri / Var(yi) * (d\mui/d\etai))^{2}] * xixi' =
(d\mui/d\etai)^{2} / Var(yi) * xixi'} Considering the whole sample, this can be
expressed in matrix form as \eqn{X'WX}. The output of this function is the
inverse of the diagonal element corresponding to the treatment estimate.
}
\keyword{internal}
