% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SandwichLayerVariance.R
\name{.get_appinv_atp}
\alias{.get_appinv_atp}
\title{(Internal) Product of \eqn{A_{pp}^{-1} A_{\tau p}^T}}
\usage{
.get_appinv_atp(x, ...)
}
\arguments{
\item{x}{a fitted \code{teeMod} model}
}
\value{
An \eqn{s\times k} matrix \eqn{A_{pp}^{-1} A_{\tau p}^T}
}
\description{
(Internal) Product of \eqn{A_{pp}^{-1} A_{\tau p}^T}
}
\keyword{internal}
