% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SandwichLayerVariance.R
\name{.get_b11}
\alias{.get_b11}
\title{(Internal) Get the B11 block of the sandwich variance estimator}
\usage{
.get_b11(x, ...)
}
\arguments{
\item{x}{A DirectAdjusted model object}

\item{...}{Arguments to be passed to sandwich::meatCL}
}
\value{
A (p+1)x(p+1) matrix the dimensions are given by the number of
terms in the covariance model (p) and an Intercept term
}
\description{
(Internal) Get the B11 block of the sandwich variance estimator
}
\details{
This is the block of the sandwich variance estimator corresponding to
the variance-covariance matrix of the covariance model coefficient estimates.
The estimates returned here are potentially clustered (by the clustering in the
experimental design) if rows in the covariance model data also exist in the
design. If there is no overlap between the two datasets, the variance-covariance
matrix is estimated assuming the observations are independent.
}
\keyword{internal}
