% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SandwichLayerVariance.R
\name{.get_b12}
\alias{.get_b12}
\title{(Internal) Get the B12 block of the sandwich variance estimator}
\usage{
.get_b12(x)
}
\arguments{
\item{x}{A DirectAdjusted model object}
}
\value{
A pxk matrix where p is the column count in the covariance model
matrix and k is the column count in the treatment model matrix
}
\description{
(Internal) Get the B12 block of the sandwich variance estimator
}
\details{
This block is the covariance matrix of the cluster-level estimating
equations for the covariance and direct adjustment models. It has a row for
each term in the covariance model and a column for each term in the direct
adjustment model. For any row that does not appear in both the experimental
design and the covariance model data, its contribution to this matrix will be
0. Thus, if there is no overlap between the two datasets, this will return a
matrix of 0's.
}
\keyword{internal}
