% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SandwichLayerVariance.R
\name{.get_b22}
\alias{.get_b22}
\title{(Internal) Get the B22 block of the sandwich variance estimator}
\usage{
.get_b22(x)
}
\arguments{
\item{x}{A DirectAdjusted model object}
}
\value{
A (p+1)x(p+1) matrix where the dimensions are given by the number of
terms in the treatment model (p) and an Intercept term. Typically, this should
be a 2x2 matrix given the dichotomous handling of treatment variables in this
package and the use of the covariance model to offer the covariance adjustment.
}
\description{
(Internal) Get the B22 block of the sandwich variance estimator
}
\details{
This block refers to the treatment assignment-level variance estimate
of the treatment effect. The \code{stats} package offers family objects with
canonical link functions, so the log-likelihood for a generalized linear model
can be written in terms of the linear predictor as \deqn{L(y_{i}, \beta, \phi,
w_{i}) = w_{i}(y_{i}\beta'x_{i} - b(\beta'x_{i})) / \phi + h(y_{i}; \phi)}
The estimating equations \eqn{\psi_{i}} given by the score function can then be
expressed as \deqn{\psi_{i} = w_{i}(y_{i} - \mu(\beta'x_{i}))x_{i} / \phi}

Where J is the number of units at the level where the treatment was assigned
in the experimental design, a matrix C of dimension nxJ is formed to indicate
which unit each subject in the design belongs to. The treatment assignment-
level estimating equations are then obtained via \eqn{C'\psi}, where \eqn{\psi}
is the matrix of estimating equations at the subject level.
}
\keyword{internal}
