% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DesignAccessors.R
\name{.get_col_from_new_data}
\alias{.get_col_from_new_data}
\title{(Internal) Extract specified \code{type} from new data set}
\usage{
.get_col_from_new_data(design, newdata, type, by = NULL, ...)
}
\arguments{
\item{design}{A \code{Design}}

\item{newdata}{A \code{data.frame}, which may or may not be the one which
was used to create \code{design}. It must have the units of assignment
variable(s) (though \code{by=} argument can be used if the name differ),
and will appropriately merge with the \code{design} the blocks, treatment
or forcings.}

\item{type}{One of "t", "f", or "b".}

\item{by}{optional; named vector or list connecting names of unit of
assignment/unitid/cluster variables in \code{design} to unit of
assignment/unitid/cluster variables in \code{data}. Names represent
variables in the Design; values represent variables in the data. Only
needed if variable names differ.}

\item{...}{Additional arguments to \code{merge()}.}
}
\value{
The column(s) belonging to the requested \code{type} in
}
\description{
(Internal) Extract specified \code{type} from new data set
}
\keyword{internal}
