% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SandwichLayerVariance.R
\name{.get_DB_covadj_bread}
\alias{.get_DB_covadj_bread}
\title{(Internal) Bread matrix of specification-based variance}
\usage{
.get_DB_covadj_bread(x, ...)
}
\arguments{
\item{x}{a fitted \code{teeMod} model}
}
\value{
a list of bread matrices
}
\description{
(Internal) Bread matrix of specification-based variance
}
\details{
Calculate bread matrix for specification-based variance estimate for
\code{teeMod} models with covariance adjustment and without absorbed effects
}
\keyword{internal}
