% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SandwichLayerVariance.R
\name{.get_DB_covadj_meat}
\alias{.get_DB_covadj_meat}
\title{(Internal) Meat matrix of specification-based variance}
\usage{
.get_DB_covadj_meat(x, ...)
}
\arguments{
\item{x}{a fitted \code{teeMod} model}
}
\value{
a list of meat matrix bounds
}
\description{
(Internal) Meat matrix of specification-based variance
}
\details{
Calculate upper and lower bound estimates of meat matrix for
specification-based variance estimate for \code{teeMod} models with
covariance adjustment and without absorbed effects
}
\keyword{internal}
