% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SandwichLayerVariance.R
\name{.get_DB_covadj_se}
\alias{.get_DB_covadj_se}
\title{(Internal) Design-based variance for models with
covariance adjustment}
\usage{
.get_DB_covadj_se(x, ...)
}
\arguments{
\item{x}{a fitted \code{teeMod} model}
}
\value{
design-based variance estimate of the main treatment effect
estimate
}
\description{
(Internal) Design-based variance for models with
covariance adjustment
}
\details{
Calculate design-based variance estimate for \code{teeMod}
models with covariance adjustment and without absorbed effects
}
\keyword{internal}
