% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_design.R
\name{.get_design}
\alias{.get_design}
\title{(Internal) Locate a \code{Design} in the call stack}
\usage{
.get_design(NULL_on_error = FALSE)
}
\arguments{
\item{NULL_on_error}{if \code{TRUE}, returns \code{NULL} if a \code{Design}
object is not found rather than throwing an error.}
}
\value{
A \code{Design}, or \code{NULL} if \code{NULL_on_error} is
\code{TRUE} and the \code{Design} can't be found.
}
\description{
\code{assigned()}/\code{ate()}/\code{ett()}/\code{cov_adj()} all need the
\code{Design} to operate. If any are called in the model without a
\code{design=} argument, this function sees if it can find the \code{Design}
in another of these functions.
}
\details{
Note that it will never look inside \code{assigned()} (gets complicated in
formulas), only in weights or \code{cov_adj()}. E.g.

\code{lm(y ~ assigned(), weights = ate(des), offest = cov_adj(mod1))}

\code{lm(y ~ assigned(), weights = ate(), offest = cov_adj(mod1, design = des))}

will both work, but

\code{lm(y ~ assigned(des), weights = ate(), offest = cov_adj(mod1))}

will fail.
}
\keyword{internal}
