% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dichotomy.R
\name{.get_dichotomy}
\alias{.get_dichotomy}
\title{(Internal) Find and validate a \code{dichotomy}}
\usage{
.get_dichotomy(dichotomy = NULL)
}
\arguments{
\item{dichotomy}{a formula or call evaluating to a formula}
}
\value{
A formula or a call that evaluates to a formula, or NULL if no
\code{dichotomy} is provided nor found in the call stack
}
\description{
(Internal) Find and validate a \code{dichotomy}
}
\details{
Whether a \code{dichotomy} is provided or not, \code{.get_dichotomy()}
finds any \code{dichotomy} argument passed to an \code{lmitt.formula()} call
up the stack or called in a \code{weights} argument. This found argument is
either used itself or used to validate a provided \code{dichotomy}.
}
\keyword{internal}
