% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SandwichLayerVariance.R
\name{.get_phi_tilde}
\alias{.get_phi_tilde}
\title{(Internal) Calculate grave\{phi\}}
\usage{
.get_phi_tilde(x, ...)
}
\arguments{
\item{x}{a fitted \code{teeMod} model}
}
\description{
(Internal) Calculate grave\{phi\}
}
\keyword{internal}
