% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{.join_design_weights}
\alias{.join_design_weights}
\title{(Internal) Expand unit of assignment level weights to the level of
the data}
\usage{
.join_design_weights(weights, design, target, data)
}
\arguments{
\item{weights}{a vector of weights sorted according to the \code{Design}}

\item{design}{a \code{Design}}

\item{target}{One of "ate" or "ett"}

\item{data}{New data}
}
\value{
a \code{WeightedDesign}
}
\description{
Helper function called during creation of the weights via \code{ate()} or
\code{ett()}
}
\keyword{internal}
