% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights_internal.R
\name{.join_spec_weights}
\alias{.join_spec_weights}
\title{(Internal) Expand unit of assignment level weights to the level of
the data}
\usage{
.join_spec_weights(
  weights,
  specification,
  target,
  weightAlias,
  data,
  dichotomy
)
}
\arguments{
\item{weights}{a vector of weights sorted according to the
\code{StudySpecification}}

\item{specification}{a \code{StudySpecification}}

\item{target}{One of "ate" or "ett"}

\item{weightAlias}{Any currently supported alias}

\item{data}{New data}

\item{dichotomy}{formula used to specify a dichotomy of a non-binary
treatment variable. The output \code{WeightedStudySpecification} object
will store this as its \code{dichotomy} slot, unless it is NULL, in which
case it will be translated to an empty \code{formula}.}
}
\value{
a \code{WeightedStudySpecification}
}
\description{
Helper function called during creation of the weights via \code{ate()} or
\code{ett()}
}
\keyword{internal}
