% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SandwichLayerVariance.R
\name{.merge_block_id_cols}
\alias{.merge_block_id_cols}
\title{(Internal) Merge multiple block IDs}
\usage{
.merge_block_id_cols(df, ids)
}
\arguments{
\item{df}{a data frame}

\item{ids}{a vector of block IDs, column names of df}
}
\value{
a data frame with a column that contains unique block number IDs
}
\description{
(Internal) Merge multiple block IDs
}
\details{
merge multiple block ID columns by the value combinations and store
the new block ID in the column \code{ids[1]}
}
\keyword{internal}
