% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_preserve_order.R
\name{.merge_preserve_order}
\alias{.merge_preserve_order}
\title{(Internal) Merge \code{data.frame}s ensuring order of first
\code{data.frame} is maintained}
\usage{
.merge_preserve_order(x, ...)
}
\arguments{
\item{x}{\code{data.frame} whose ordering is to be maintained}

\item{...}{Additional arguments to \code{merge()}, particularly a second
\code{data.frame} and a \code{by=} argument.}
}
\value{
Merged \code{data.frame} with the same ordering as \code{x}.
}
\description{
(Internal) Merge \code{data.frame}s ensuring order of first
\code{data.frame} is maintained
}
\keyword{internal}
