% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design.R
\name{.new_Design}
\alias{.new_Design}
\title{Create a new \code{Design} object.}
\usage{
.new_Design(
  form,
  data,
  type,
  subset = NULL,
  call = NULL,
  dichotomy = stats::formula()
)
}
\arguments{
\item{form}{Formula to create Design, see help for \code{rcr_design()},
\code{rd_design()} or \code{obs_design()} for details.}

\item{data}{The data set}

\item{type}{One of "RCT", "RD", or "Obs"}

\item{subset}{Any subset information}

\item{call}{The call generating the \code{Design}.}

\item{dichotomy}{If present, the dichotomization formula}
}
\value{
A new Design object
}
\description{
Create a new \code{Design} object.
}
\keyword{internal}
