% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StudySpecification.R
\name{.new_StudySpecification}
\alias{.new_StudySpecification}
\title{(Internal) Create a new \code{StudySpecification} object.}
\usage{
.new_StudySpecification(
  form,
  data,
  type,
  subset = NULL,
  call = NULL,
  na.fail = TRUE
)
}
\arguments{
\item{form}{Formula to create StudySpecification, see help for
\code{rcr_spec()}, \code{rd_spec()} or \code{obs_spec()} for details.}

\item{data}{The data set}

\item{type}{One of "RCT", "RD", or "Obs"}

\item{subset}{Any subset information}

\item{call}{The call generating the \code{StudySpecification}.}

\item{na.fail}{Should it error on NA's (\code{TRUE}) or remove them
(\code{FALSE})?}
}
\value{
A new StudySpecification object
}
\description{
Helper function to create a new \code{StudySpecification}. Called internally
from \code{rct_spec()}, \code{rd_spec()} or \code{obs_spec()}.
}
\keyword{internal}
