% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DirectAdjusted.R
\name{.order_samples}
\alias{.order_samples}
\title{Order observations used to fit a \code{DirectAdjusted} model and its
covariance adjustment model}
\usage{
.order_samples(x, by = NULL, ...)
}
\arguments{
\item{x}{a fitted \code{DirectAdjusted} object}

\item{by}{character vector; indicates unit of assignment columns to generate
ID's from; default is NULL, which uses the unit of assignment columns specified
in the \code{DirectAdjusted} object's \code{Design} slot}

\item{...}{arguments passed to methods}
}
\value{
a list of four named vectors
}
\description{
Order observations used to fit a \code{DirectAdjusted} model and its
covariance adjustment model
}
\details{
\code{.order_samples} underpins the ordering for \code{.make_uoa_ids}
and \code{estfun.DirectAdjusted}. This function needs to order the rows in
in \eqn{\mathcal{Q}\cup\mathcal{C}}, but it also needs to explain how the
original matrices of estimating equations should be indexed so the contributions
from both sets of equations match. So instead of returning a numeric vector,
which one might expect for an ordering function, \code{.order_samples}
returns a list of vectors, which is explained in the Return section. Ultimately,
the order is given by concatenating the vectors stored in \code{Q_not_C},
\code{Q_in_C}, and \code{C_not_q} (\code{Q_in_C} and \code{C_in_Q} are
interchangeable in terms of deriving the order). The names of the
\code{Q_not_C} and \code{Q_in_C} vectors correspond to row indices of the
matrix of estimating equations for the ITT effect model, while the names of
the \code{C_in_Q} and \code{C_not_Q} vectors correspond to row indices of
the matrix of estimating equations for the covariance adjustment model.\n\n
When a \code{by} argument is provided to \code{cov_adj}, it is used to
deduce the order.
}
\keyword{internal}
