% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SandwichLayerVariance.R
\name{.prepare_design_matrix}
\alias{.prepare_design_matrix}
\title{(Internal) Helper function for design-based meat matrix calculation}
\usage{
.prepare_design_matrix(x, ...)
}
\arguments{
\item{x}{a fitted \code{teeMod} model}
}
\description{
(Internal) Helper function for design-based meat matrix calculation
}
\keyword{internal}
