% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DesignConverters.R
\name{.remove_forcing}
\alias{.remove_forcing}
\title{(Internal) Removes the forcing column entirely from a \code{Design}}
\usage{
.remove_forcing(des)
}
\arguments{
\item{des}{A \code{Design}}
}
\value{
The \code{Design} without any forcing variable
}
\description{
In preparation for converting an RD \code{Design} to another \code{Design},
this will strip the forcing variable entirely. It is removed from the data
(both \code{@structure} and \code{@column_index}), as well as from the
formula stored in \code{@call}.
}
\details{
Note that the output \code{Design} will fail a validity check (with
\code{validObject()}) due to an RD \code{Design} requiring a forcing
variable, so change the \code{@type} immediately.
}
\keyword{internal}
