% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SandwichLayer.R
\name{.sanitize_C_uoas}
\alias{.sanitize_C_uoas}
\title{Generate a list of sanitized units of assignment from C}
\usage{
.sanitize_C_uoas(x, cluster = NULL, verbose = TRUE, ...)
}
\arguments{
\item{x}{a \code{SandwichLayer} object.}

\item{cluster}{Defaults to NULL, which means unit of assignment columns
indicated in the Design will be used to generate clustered covariance estimates.
A non-NULL argument to \code{cluster} specifies a string or character vector of
column names appearing in both the covariance adjustment and quasiexperimental
samples that should be used for clustering covariance estimates.}

\item{verbose}{Boolean defaulting to TRUE, which will produce rather than
swallow any warnings about the coding of the units of assignment in the
covariance adjustment model data}

\item{...}{arguments passed to methods}
}
\value{
A vector of length \eqn{|C|}, where C is the covariance adjustment sample
}
\description{
Generate a list of sanitized units of assignment from C
}
\keyword{internal}
