% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DirectAdjusted.R
\name{.sanitize_Q_ids}
\alias{.sanitize_Q_ids}
\title{Return ID's for observations in the quasiexperimental sample Q}
\usage{
.sanitize_Q_ids(x, id_col = NULL, ...)
}
\arguments{
\item{x}{a fitted \code{DirectAdjusted} object}

\item{id_col}{character vector or list; optional. Specifies column names that appear in
both the covariance adjustment dataframe C and quasiexperimental dataframe Q.
Defaults to NULL, in which case unit of assignment columns indicated in the
Design will be used to generate ID's.}

\item{...}{arguments passed to methods}
}
\value{
A vector of length \eqn{|Q|}, where Q is the quasiexperimental sample.
}
\description{
Return ID's for observations in the quasiexperimental sample Q
}
\keyword{internal}
