% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DirectAdjusted.R
\name{.sanitize_Q_uoas}
\alias{.sanitize_Q_uoas}
\title{Generate a list of sanitized units of assignment from Q}
\usage{
.sanitize_Q_uoas(x, cluster = NULL, ...)
}
\arguments{
\item{x}{a fitted \code{DirectAdjusted} object}

\item{cluster}{Defaults to NULL, which means unit of assignment columns
indicated in the Design will be used to generate clustered covariance estimates.
A non-NULL argument to \code{cluster} specifies a string or character vector of
column names appearing in both the covariance adjustment and quasiexperimental
samples that should be used for clustering covariance estimates.}

\item{...}{arguments passed to methods}
}
\value{
A vector of length \eqn{|Q|}, where Q is the quasiexperimental sample
}
\description{
Generate a list of sanitized units of assignment from Q
}
\keyword{internal}
