% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DirectAdjusted.R
\name{.sanitize_uoas}
\alias{.sanitize_uoas}
\title{Generate a vector of sanitized units of assignment from C and Q}
\usage{
.sanitize_uoas(x, cluster = NULL, ca = x$model$`(offset)`, verbose = TRUE, ...)
}
\arguments{
\item{x}{a fitted \code{DirectAdjusted} object}

\item{cluster}{Defaults to NULL, which means unit of assignment columns
indicated in the Design will be used to generate clustered covariance estimates.
A non-NULL argument to \code{cluster} specifies a string or character vector of
column names appearing in both the covariance adjustment and quasiexperimental
samples that should be used for clustering covariance estimates.}

\item{ca}{SandwichLayer object storing information about the covariance
adjustment model; usually stored as the \code{offset} of a \code{DirectAdjusted}
object when covariance adjustment is performed}

\item{verbose}{Boolean defaulting to TRUE, which will produce rather than
swallow any warnings about the coding of the units of assignment in the
covariance adjustment model data}

\item{...}{arguments passed to methods}
}
\value{
A named vector of length \eqn{|Q| + |C \ Q|}, where Q and C represent
the sets of observations in the quasiexperimental sample and covariance
adjustment sample, respectively. Values can be "Q", for observations that only
appear in Q, "C", for those that only appear in C, and "QC" for those that
appear in both. Names correspond to the unit of assignment ID.
}
\description{
Generate a vector of sanitized units of assignment from C and Q
}
\keyword{internal}
