% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SandwichLayer.R
\name{.show_layer}
\alias{.show_layer}
\title{(Internal) \code{show} helper for
\code{PreSandwichLayer}/\code{SandwichLayer}}
\usage{
.show_layer(object)
}
\arguments{
\item{object}{\code{PreSandwichLayer} or \code{SandwichLayer}}
}
\value{
\code{object}, invisibly
}
\description{
(Internal) \code{show} helper for
\code{PreSandwichLayer}/\code{SandwichLayer}
}
\keyword{internal}
