% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DirectAdjusted.R
\name{.txt_fn}
\alias{.txt_fn}
\title{Treatment variable name}
\usage{
.txt_fn(object)
}
\arguments{
\item{object}{\code{DirectAdjusted} object}
}
\value{
Character string identifying the name (e.g "assigned()" or
"a.(des)")
}
\description{
(Internal) Obtain which variaiton of \code{assigned()}, \code{a.()} or
\code{z.()} is used in the model
}
\keyword{internal}
