% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_by.R
\name{.update_by}
\alias{.update_by}
\title{(Internal) Use \code{by} to update \code{Design} with new variable
names}
\usage{
.update_by(design, data, by)
}
\arguments{
\item{design}{A \code{Design}}

\item{data}{\code{Data set}}

\item{by}{named vector or list connecting names of cluster/unit of
assignment variables in \code{design} to cluster/unit of assignment
variables in \code{data}. Names represent variables in the Design; values
represent variables in the data.}
}
\value{
A \code{Design} with updated variable names
}
\description{
Helper function used to update the variable names of a \code{Design} when
user passes a \code{by=} argument to align variable names between data sets.
}
\keyword{internal}
