% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_adj.R
\name{.update_ca_model_formula}
\alias{.update_ca_model_formula}
\title{(Internal) Add columns for merging covariance adjustment and direct
adjustment samples to model formula}
\usage{
.update_ca_model_formula(model, by = NULL, specification = NULL)
}
\arguments{
\item{model}{any model that inherits from a \code{glm}, \code{lm}, or
\code{robustbase::lmrob} object.}

\item{by}{named vector; optional.}

\item{specification}{a \code{StudySpecification} object; optional. If not
provided, it can be retrieved from the call stack if passed to suitable
calls.}
}
\value{
formula
}
\description{
(Internal) Add columns for merging covariance adjustment and direct
adjustment samples to model formula
}
\details{
This function is typically used prior to
\code{.get_data_from_model()} and incorporates information provided in a
\code{by} argument to ensure the necessary columns for merging the two
samples are included in any \code{model.frame()} calls.
}
\keyword{internal}
