% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DesignAccessors.R
\name{.update_call_formula}
\alias{.update_call_formula}
\title{(Internal) Updates \code{des@call}'s formula with the currently defined
variable names.}
\usage{
.update_call_formula(design)
}
\arguments{
\item{design}{A \code{Design}}
}
\value{
An updated formula
}
\description{
Helper function to update the \code{call} with the appropriate variable
names after they've been modified. Called within \code{Design} replacers.
}
\details{
It's return should be stuck into the design via \code{des@call$formula <-
.update_call_formula(des)}
}
\keyword{internal}
