% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StudySpecificationUtilities.R
\name{.update_form_to_unit_of_assignment}
\alias{.update_form_to_unit_of_assignment}
\title{(Internal) Rename cluster/unitid/uoa in a formula to
unit_of_assignment for internal consistency}
\usage{
.update_form_to_unit_of_assignment(form)
}
\arguments{
\item{form}{A formula passed to \code{.new_StudySpecification()}}
}
\value{
The formula with "cluster"/"unitid"/"uoa" replace with
"unit_of_assignment"
}
\description{
Internally, we always refer to uoa/cluster/unitid as "unit_of_assignment"
}
\keyword{internal}
