% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dichotomy.R
\name{.validate_dichotomy}
\alias{.validate_dichotomy}
\title{(Internal) Validate a dichotomy against other dichotomies found in the
call stack}
\usage{
.validate_dichotomy(possible_dichotomy)
}
\arguments{
\item{possible_dichotomy}{a list or a formula, the former coming from
\code{.find_dichotomies()}}
}
\value{
formula for a dichotomy or NULL
}
\description{
(Internal) Validate a dichotomy against other dichotomies found in the
call stack
}
\keyword{internal}
