% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{.weights_calc}
\alias{.weights_calc}
\title{(Internal) Worker function for weight calculation}
\usage{
.weights_calc(design, target, dichotomy, by, data)
}
\arguments{
\item{design}{a \code{Design} object created by one of \code{rct_design()},
\code{rd_design()}, or \code{obs_design()}.}

\item{target}{One of "ate" or "ett"; \code{ate()} and \code{ett()} chooses
these automatically.}

\item{dichotomy}{optional; a formula defining the dichotomy of the treatment
variable if it isn't already \code{0}/\code{1}. See details of help for
\code{rct_design()} e.g. for details.}

\item{by}{optional; named vector or list connecting names of cluster/unit of
assignment variables in \code{design} to cluster/unit of assignment
variables in \code{data}. Names represent variables in the Design; values
represent variables in the data. Only needed if variable names differ.}

\item{data}{optionally the data for the analysis to be performed on. May be
excluded if these functions are included as the \code{weights} argument of
a model.}
}
\value{
a \code{WeightedDesign} object
}
\description{
Called from \code{ate()} or \code{ett()}.
}
\author{
Josh Errickson
}
