% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DirectAdjusted.R
\name{estfun.DirectAdjusted}
\alias{estfun.DirectAdjusted}
\title{Extract Empirical Estimating Functions}
\usage{
\method{estfun}{DirectAdjusted}(object, ...)
}
\arguments{
\item{object}{DirectAdjusted}
}
\value{
n x (k+1) matrix corresponding to the empirical estimating equations
of the ITT model incorporating covariance adjustment. Here, n represents the
total number of units in the covariance adjustment model and ITT effect model
samples combined, and k represents the number of treatments. The output
matrix is ordered such that the units used to fit the ITT effect model
comprise the intial rows, and those solely used to fit the covariance
adjustment model follow.
}
\description{
Extract Empirical Estimating Functions
}
