% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SandwichLayerVariance.R
\name{estfun.lmrob}
\alias{estfun.lmrob}
\title{Estfun method for lmrob objects}
\usage{
\method{estfun}{lmrob}(x, ...)
}
\arguments{
\item{x}{An \code{lmrob} object produced using an MM/SM estimator chain}

\item{...}{Additional arguments to be passed to \code{estfun}}
}
\value{
A \eqn{n\times }(p+1) matrix where the first column corresponds to
the scale estimate and the remaining \eqn{p} colums correspond to the
coefficients
}
\description{
Estfun method for lmrob objects
}
\author{
lrd author 2
}
