% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terms.lmitt.formula.R
\name{formula.lmitt.terms}
\alias{formula.lmitt.terms}
\title{Return the \code{lmitt.formula} associated with a \code{lmitt.terms} object}
\usage{
\method{formula}{lmitt.terms}(x, ...)
}
\arguments{
\item{x}{\code{lmitt.terms}.}

\item{...}{Arguments passed to \code{...} do not modify the
\code{lmitt.terms} object given.}
}
\value{
The formula associated with the \code{lmitt.terms} object; in other
words, the \code{lmitt.formula} it was created from (as a standard
\code{formula} class), rather than the \code{model_form} attribute.
}
\description{
Return the \code{lmitt.formula} associated with a \code{lmitt.terms} object
}
\examples{
data(simdata)
spec <- rct_spec(z ~ 1, simdata)
form <- build_lmitt_formula("y", "1", rhstype = "intercept", absorb = FALSE)
tt <- terms(form, spec)
stats::formula(tt)
}
