% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SandwichLayerVariance.R
\name{get_da_bread}
\alias{get_da_bread}
\title{Get the "bread" matrix for the Direct Adjustment Model}
\usage{
get_da_bread(x)
}
\arguments{
\item{x}{DirectAdjusted}
}
\value{
1x1 matrix. We follow Carroll et al.'s formulation (2006, pp. 373) of
this matrix, which is the sum of the gradients of the direct adjustment
model's estimating equations taken with respect to the treatment variable.
}
\description{
Get the "bread" matrix for the Direct Adjustment Model
}
