% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SandwichLayerVariance.R
\name{get_overlap_vcov_matrix}
\alias{get_overlap_vcov_matrix}
\title{Generate Cluster-level Estimating Equations Covariance Matrix
for Overlapping Rows in Experimental and Covariance Model Data}
\usage{
get_overlap_vcov_matrix(x)
}
\arguments{
\item{x}{DirectAdjusted}
}
\value{
matrix of dim (p+1) x t where p is the number of covariates in the
covariance model and t is the number of treatment levels
}
\description{
\code{get_overlap_vcov_matrix} returns the covariance matrix of the
cluster-level estimating equations for rows that appear in both the experimental
and covariance model data
}
