% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DesignStructure.R
\name{get_structure}
\alias{get_structure}
\title{Returns \code{Design} Structure Information}
\usage{
get_structure(design, binary = FALSE)
}
\arguments{
\item{design}{A \code{Design}.}

\item{binary}{Binary, default \code{FALSE}. If \code{TRUE} and the design
contains a \code{dichotomy}, replace the treatment column with its binary
representation. Has no effect if \code{design} is not dichotomized.}
}
\value{
A \code{data.frame} containing the structure of the \code{design}.
}
\description{
Creates a \code{data.frame} containing the information at the unit of
assignment/cluster level.
}
