% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DesignUtilities.R
\name{has_binary_treatment}
\alias{has_binary_treatment}
\title{Check whether treatment stored in a \code{Design} object is binary}
\usage{
has_binary_treatment(des)
}
\arguments{
\item{des}{\code{Design} object}
}
\value{
logical vector of length 1
}
\description{
Check whether treatment stored in a \code{Design} object is binary
}
