% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DesignUtilities.R
\name{identical_Designs}
\alias{identical_Designs}
\title{Test equality of two \code{Design} objects}
\usage{
identical_Designs(x, y)
}
\arguments{
\item{x}{A \code{Design} object.}

\item{y}{A \code{Design} object.}
}
\value{
Logical, are \code{x} and \code{y} identical?
}
\description{
Check whether two \code{Design} objects are identical.
}
