% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StudySpecificationUtilities.R
\name{identical_StudySpecifications}
\alias{identical_StudySpecifications}
\title{Test equality of two \code{StudySpecification} objects}
\usage{
identical_StudySpecifications(x, y)
}
\arguments{
\item{x}{A \code{StudySpecification} object.}

\item{y}{A \code{StudySpecification} object.}
}
\value{
Logical, are \code{x} and \code{y} identical?
}
\description{
Check whether two \code{StudySpecification} objects are
identical.
}
\examples{
data(schooldata)
sdata <- schooldata
spec1 <- rct_spec(treatment ~ unitid(schoolid) + block(state), sdata)
sdata$treatment <- sample(sdata$treatment)
spec2 <- rct_spec(treatment ~ unitid(schoolid) + block(state), sdata)
identical_StudySpecifications(spec1, spec2)
sdata$treatment <- schooldata$treatment
spec3 <- rct_spec(treatment ~ unitid(schoolid) + block(state), sdata)
identical_StudySpecifications(spec1, spec3)
}
