% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DesignUtilities.R
\name{is_dichotomized}
\alias{is_dichotomized}
\title{Check if Design's treatment was dichotomized}
\usage{
is_dichotomized(des)
}
\arguments{
\item{des}{A design}
}
\value{
TRUE if \code{des} was dichotomized; FALSE otherwise
}
\description{
Check if Design's treatment was dichotomized
}
