% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ittestimate.R
\name{ittestimate}
\alias{ittestimate}
\title{Estimate Treatment Effect}
\usage{
ittestimate(
  design,
  data,
  outcome,
  target = "ate",
  covAdjModel = NULL,
  unitOfAssignmentIds = NULL,
  ...,
  weights = NULL
)
}
\arguments{
\item{design}{study Design}

\item{data}{Data for analysis}

\item{outcome}{string containing name of outcome variable in \code{data}}

\item{target}{a function returning a WeightedDesign object, or either "ate"
(default) or "ett"}

\item{covAdjModel}{optional; covariable adjustment model}

\item{unitOfAssignmentIds}{optional; list connecting names of units of
assignment/clusters/units variables in \code{design} to units of
assignment/clusters/units variables in \code{data}}

\item{...}{Additional arguments to the model.}

\item{weights}{optional; manually include weights. If included, weights will
\strong{not} be automatically generated. Instead, the \code{weights} argument should
include the product of any externally created weights and the result of
\code{ate()} or \code{ett()}. For example, \code{weights = myweights*ate(des)}.}
}
\value{
Estimated treatment effect
}
\description{
Estimate Treatment Effect
}
