% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmitt.formula.R
\docType{class}
\name{lmitt.formula-class}
\alias{lmitt.formula-class}
\title{Formulas generated from inputs to \code{lmitt()}}
\description{
The \code{terms()} method for this class returns a \code{lmitt.terms}
object that helps generate artifacts for fitting \code{teeMod} objects
}
\keyword{internal}
