% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmitt.R
\name{lmitt}
\alias{lmitt}
\title{Linear Model for Intention To Treat}
\usage{
lmitt(
  formula,
  design = NULL,
  target = NULL,
  data,
  subset,
  weights,
  na.action,
  method = "qr",
  model = TRUE,
  x = FALSE,
  y = FALSE,
  qr = TRUE,
  singular.ok = TRUE,
  contrasts = NULL,
  offset,
  ...
)
}
\arguments{
\item{formula}{See \code{lm()}}

\item{design}{Optional, explicitly specify the \code{Design} to be used. If
the \code{Design} is specified elsewhere in the model (e.g. passed as an
argument to any of \code{ate()}, \code{ett()}, \code{cov_adj()} or
\code{adopters()}) it will be found automatically and does not need to be
passed here as well. (If the \code{Design} is found in the model, this
argument is ignored.)}

\item{target}{Optional, explicitly specify the estimand. If the
\code{weights} are generated using either \code{ate()} or \code{ett()},
the \code{target]} will be found automatically. Otherwise, specify whether
the goal is estimating ATE ("ate") or ETT ("ett"). (If weights are
specified, this argument is ignored.)}

\item{data}{See \code{lm()}}

\item{subset}{See \code{lm()}}

\item{weights}{See \code{lm()}}

\item{na.action}{See \code{lm()}}

\item{method}{See \code{lm()}}

\item{model}{See \code{lm()}}

\item{x}{See \code{lm()}}

\item{y}{See \code{lm()}}

\item{qr}{See \code{lm()}}

\item{singular.ok}{See \code{lm()}}

\item{contrasts}{See \code{lm()}}

\item{offset}{See \code{lm()}}

\item{...}{See \code{lm()}}
}
\value{
\code{DirectAdjusted} model.
}
\description{
Linear Model for Intention To Treat
}
