% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teeMod.model.frame.R
\docType{class}
\name{lmitt.terms-class}
\alias{lmitt.terms-class}
\title{teeMod.model.frame class
Class for lmitt formulas}
\description{
teeMod.model.frame class
Class for lmitt formulas
}
\details{
This has a specific \code{model.matrix()} method that interprets our
dichotomy and absorption specifications
}
