% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbph.R
\name{lmitt_pbph}
\alias{lmitt_pbph}
\title{Peters-Belson with Prognostic Heterogeneity}
\usage{
lmitt_pbph(mod1, design, data)
}
\arguments{
\item{mod1}{First stage fitted model.}

\item{design}{\code{Design} object}

\item{data}{Data where variables in \code{form} live.}
}
\value{
A \code{pbph} object which extends \code{lm}. Can be
passed to \code{summary} or \code{confint}.

The return contains an additional object, \code{pbph}, which \code{lm}
doesn't include. \code{pbph} contains \code{mod1}, a copy of the first
stage model, and \code{data}, which includes the \code{data} augmented
with \code{treatment} and \code{predicted}.
}
\description{
Performs Peters-Belson with Prognostic Heterogeneity on the data.
}
