% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmrob_methods.R
\name{estfun.lmrob}
\alias{estfun.lmrob}
\alias{bread.lmrob}
\title{Generate matrix of estimating equations for \code{lmrob()} fit}
\usage{
\method{estfun}{lmrob}(x, ...)

\method{bread}{lmrob}(x, ...)
}
\arguments{
\item{x}{An \code{lmrob} object produced using an MM/SM estimator chain}

\item{...}{Additional arguments to be passed to \code{bread}}
}
\value{
A \eqn{n\times }(p+1) matrix where the first column corresponds to
the scale estimate and the remaining \eqn{p} colums correspond to the
coefficients

A \eqn{p\times }(p+1) matrix where the first column corresponds to
the scale estimate and the remaining \eqn{p} colums correspond to the
coefficients
}
\description{
Generate matrix of estimating equations for \code{lmrob()} fit

Extract bread matrix from an \code{lmrob()} fit
}
\details{
This is part of a workaround for an issue in the robustbase code
affecting sandwich covariance estimation. The issue in question is issue
#6471, robustbase project on R-Forge. This function contributes to
providing sandwich estimates of covariance-adjusted standard errors for
robust linear covariance adjustment models.

This is part of a workaround for an issue in the robustbase code
affecting sandwich covariance estimation. The issue in question is issue
#6471, robustbase project on R-Forge. This function contributes to
providing sandwich estimates of covariance-adjusted standard errors for
robust linear covariance adjustment models.
}
\author{
Ben B. Hansen
}
