% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{michigan_school_pairs}
\alias{michigan_school_pairs}
\title{Intervention data from a pair-matched study of schools in Michigan}
\format{
A \code{data.frame} with 14 rows and 13 columns.
\itemize{
\item{schoolid} school id
\item{blk} block
\item{z} treatment variable
\item{male_g11_perc} percentage of G11 male students
\item{female_g11_perc} percentage of G11 female students
\item{am_g11_perc} percentage of G11 American Indian/Alaska Native students
\item{asian_g11_perc} percentage of G11 Asian students
\item{hisp_g11_perc} percentage of G11 Hispanic students
\item{black_g11_perc} percentage of G11 Black students
\item{white_g11_perc} percentage of G11 White students
\item{pacific_g11_perc} percentage of G11 Hawaiian Native/Pacific Islander students
\item{tr_g11_perc} percentage of G11 Two or More Races students
\item{g11} Number of G11 students
}
}
\usage{
michigan_school_pairs
}
\description{
Schools matched into pairs or triplets approximating the subset
of participating schools in Pane et al. (2014) from Michigan
}
\references{
Pane, John F., et al. "Effectiveness of cognitive tutor algebra I at scale."
\emph{Educational Evaluation and Policy Analysis} 36.2 (2014): 127-144.
}
\keyword{dataset}
