% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmitt.formula.model.frame.R,
%   R/model.frame.lmitt.terms.R
\name{model.frame.lmitt.terms}
\alias{model.frame.lmitt.terms}
\title{Build a model frame from a lmitt.formula}
\usage{
\method{model.frame}{lmitt.terms}(formula, data, ...)

\method{model.frame}{lmitt.terms}(formula, data, ...)
}
\arguments{
\item{formula}{\code{lmitt.terms}.}

\item{data}{dataframe.}

\item{...}{Additional arguments to inform the model frame generation, such
as weights, subset, or offset calls.}
}
\value{
dataframe with a \code{terms} attribute storing the original
\code{lmitt.terms} object
}
\description{
The model frame will reflect the formula stored in the \code{model_form}
attribute of the \code{lmitt.terms} object passed to the \code{formula}
argument. If the \code{absorb} attribute of that object is TRUE, the model
frame will have an additional column for block identifiers. The \code{terms}
attribute of the model frame is the original \code{lmitt.terms} object passed
to the \code{formula} argument.
}
\examples{
data(simdata)
spec <- rct_spec(z ~ 1, simdata)
form <- build_lmitt_formula("y", "1", rhstype = "intercept", absorb = FALSE,
                            specification = quote(spec),
                            data = quote(simdata))
tt <- terms(form, spec)
stats::model.frame(tt, simdata)

blkd_spec <- rct_spec(z ~ unitid(uoa1, uoa2) + block(bid), simdata)
form <- build_lmitt_formula("y", "1", rhstype = "intercept", absorb = TRUE,
                            specification = quote(blkd_spec),
                            data = quote(simdata))
tt <- terms(form, spec)
stats::model.frame(tt, simdata)
}
