% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.matrix.lmitt.terms.R,
%   R/teeMod.model.frame.R
\name{model.matrix.lmitt.terms}
\alias{model.matrix.lmitt.terms}
\title{Build a model matrix from a \code{lmitt.terms} object and its associated
model frame}
\usage{
\method{model.matrix}{lmitt.terms}(object, data, ...)

\method{model.matrix}{lmitt.terms}(object, data, ...)
}
\arguments{
\item{object}{\code{lmitt.terms}.}

\item{data}{model frame generated using \code{object}.}

\item{...}{Additional arguments for forming the model matrix such as
\code{xlev}. See \code{stats::model.matrix.default()}.}
}
\value{
matrix, possibly with a \code{block.means} attribute storing the
column means by block

matrix, possibly with a \code{block.means} attribute storing the
column means by block
}
\description{
The model matrix reflects the formula in the \code{model_form} attribute of
the \code{lmitt.terms} object passed to the \code{object} argument, hence the
need for the \code{data} argument to have columns from the model frame rather
than the dataframe from which the model frame may have been generated.
Aside from the intercept column, columns in the output matrix are
block-centered if the \code{absorb} attribute of \code{object} is TRUE; the
block means are stored as a \code{block.means} attribute of the model matrix
in this case.

The model matrix reflects the formula in the \code{model_form} attribute of
the \code{lmitt.terms} object passed to the \code{object} argument, hence the
need for the \code{data} argument to have columns from the model frame rather
than the dataframe from which the model frame may have been generated.
Aside from the intercept column, columns in the output matrix are
block-centered if the \code{absorb} attribute of \code{object} is TRUE; the
block means are stored as a \code{block.means} attribute of the model matrix
in this case.
}
\examples{
data(simdata)
spec <- rct_spec(z ~ 1, simdata)
form <- build_lmitt_formula("y", "1", rhstype = "intercept", absorb = FALSE,
                            specification = quote(spec),
                            data = quote(simdata))
tt <- stats::terms(form, spec)
mf <- stats::model.frame(tt, simdata)
model.matrix(tt, mf)

blkd_spec <- rct_spec(z ~ unitid(uoa1, uoa2) + block(bid), simdata)
form <- build_lmitt_formula("y", "1", rhstype = "intercept", absorb = TRUE,
                            specification = quote(blkd_spec),
                            data = quote(simdata))
tt <- stats::terms(form, spec)
mf <- stats::model.frame(tt, simdata)
stats::model.matrix(tt, mf)
data(simdata)
spec <- rct_spec(z ~ 1, simdata)
form <- build_lmitt_formula("y", "1", rhstype = "intercept", absorb = FALSE,
                            specification = quote(spec),
                            data = quote(simdata))
tt <- stats::terms(form, spec)
mf <- stats::model.frame(tt, simdata)
model.matrix(tt, mf)

blkd_spec <- rct_spec(z ~ unitid(uoa1, uoa2) + block(bid), simdata)
form <- build_lmitt_formula("y", "1", rhstype = "intercept", absorb = TRUE,
                            specification = quote(blkd_spec),
                            data = quote(simdata))
tt <- stats::terms(form, spec)
mf <- stats::model.frame(tt, simdata)
stats::model.matrix(tt, mf)
}
