% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.matrix.lmitt.terms.R
\name{model.matrix.teeMod}
\alias{model.matrix.teeMod}
\title{Return a model matrix associated with a model fit by \code{lmitt()}}
\usage{
\method{model.matrix}{teeMod}(object, ...)
}
\arguments{
\item{object}{\code{teeMod}, or a model fit by \code{lmitt()}.}

\item{...}{Additional arguments for forming the model matrix such as
\code{xlev}. See \code{stats::model.matrix.default()}.}
}
\description{
Return a model matrix associated with a model fit by \code{lmitt()}
}
\examples{
data(simdata)
spec <- rct_spec(z ~ unitid(uoa1, uoa2), simdata)
tm <- lmitt(y ~ 1, spec, simdata)
stats::model.matrix(tm)

blkd_spec <- rct_spec(z ~ unitid(uoa1, uoa2) + block(bid), simdata)
tm <- lmitt(y ~ 1, blkd_spec, simdata, absorb = TRUE)
stats::model.matrix(tm)
}
