% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design.R
\name{Obs_Design}
\alias{Obs_Design}
\title{Specify Observational Data Design}
\usage{
Obs_Design(formula, data, subset = NULL)
}
\arguments{
\item{formula}{defines the design components}

\item{data}{the data set.}

\item{subset}{optionally subset the data before creating the design object}
}
\value{
a Design object of type "Obs" for use in further analysis
}
\description{
Generates an Observationl Data Design object with the given specifications.
}
\details{
The \code{formula} must include \code{cluster()} to identify the units of assignment
(one or more variables), it may optionally contain \code{strata()} as well.
}
