% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SandwichLayer.R
\docType{class}
\name{PreSandwichLayer-class}
\alias{PreSandwichLayer-class}
\title{(Internal) model predictions with some model artifacts, as S4 object}
\description{
(Internal) model predictions with some model artifacts, as S4 object
}
\section{Slots}{

\describe{
\item{\code{.Data}}{numeric vector of predictions}

\item{\code{fitted_covariance_model}}{model standing behind the predictions}

\item{\code{prediction_gradient}}{matrix, predictions gradient w/r/t model params}
}}

\keyword{internal}
