% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SandwichLayer.R
\name{show,PreSandwichLayer-method}
\alias{show,PreSandwichLayer-method}
\title{Show a \code{PreSandwichLayer} or \code{SandwichLayer}}
\usage{
\S4method{show}{PreSandwichLayer}(object)
}
\arguments{
\item{object}{\code{PreSandwichLayer} or \code{SandwichLayer} object}
}
\value{
an invisible copy of \code{object}
}
\description{
Display information about a \code{PreSandwichLayer} or
\code{SandwichLayer} object
}
\examples{
data(schooldata)
data(studentdata)
covadj_mod <- lm(math ~ gpa + grade, studentdata[seq_len(15),])
suppressWarnings(psl <- cov_adj(covadj_mod))
show(psl)
identical(subset(psl, c(rep(TRUE, 4), rep(FALSE, 11))),
          psl[seq_len(4)])
}
