% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SandwichLayer.R
\name{subset,PreSandwichLayer-method}
\alias{subset,PreSandwichLayer-method}
\alias{[,PreSandwichLayer,ANY,ANY,ANY-method}
\title{PreSandwichLayer and SandwichLayer subsetting}
\usage{
\S4method{subset}{PreSandwichLayer}(x, subset)

\S4method{[}{PreSandwichLayer,ANY,ANY,ANY}(x, i)
}
\arguments{
\item{x}{\code{PreSandwichLayer} object}

\item{subset}{Logical vector identifying values to keep or drop}

\item{i}{indices specifying elements to extract or replace. See
\code{help("[")} for further details.}
}
\value{
\code{x} subset by \code{i}

\code{x} subset by \code{i}
}
\description{
PreSandwichLayer and SandwichLayer subsetting
}
