% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.Design.R
\name{print.summary.Design}
\alias{print.summary.Design}
\title{Print summary object}
\usage{
\method{print}{summary.Design}(x, ...)
}
\arguments{
\item{x}{Design object}

\item{...}{Other args}
}
\value{
object, invisibly
}
\description{
Print summary object
}
