% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbph.R
\name{print.summary.DirectAdjustedPBPH}
\alias{print.summary.DirectAdjustedPBPH}
\title{Print summary of \code{DirectAdjustedPBPH} object}
\usage{
\method{print}{summary.DirectAdjustedPBPH}(
  x,
  digits = max(3L, getOption("digits") - 3L),
  signif.stars = getOption("show.signif.stars"),
  ...
)
}
\arguments{
\item{x}{\code{summary.DirectAdjustedPBPH} object}

\item{digits}{the number of significant digits to use when printing.}

\item{signif.stars}{logical. If ‘TRUE’, ‘significance stars’ are printed for
each coefficient.}

\item{...}{Other args}
}
\value{
object, invisibly
}
\description{
Print summary of \code{DirectAdjustedPBPH} object
}
