% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design.R
\name{RD_Design}
\alias{RD_Design}
\title{Specify Regression Discontinuity Design}
\usage{
RD_Design(formula, data, subset = NULL)
}
\arguments{
\item{formula}{defines the design components}

\item{data}{the data set.}

\item{subset}{optionally subset the data before creating the design object}
}
\value{
a Design object of type "RD" for use in further analysis
}
\description{
Generates an RD Design object with the given specifications.
}
\details{
The \code{formula} must include \code{cluster()} to identify the units of assignment
(one or more variables), it may optionally contain \code{strata()} and/or
\code{forcing()} as well.
}
