% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SandwichLayerVariance.R
\name{.get_a22_inverse}
\alias{.get_a22_inverse}
\alias{.get_a11_inverse}
\alias{.get_a21}
\alias{.get_tilde_a22_inverse}
\alias{.get_tilde_a21}
\title{(Internal) Compute variance blocks}
\usage{
.get_a22_inverse(x, ...)

.get_a11_inverse(x)

.get_a21(x)

.get_tilde_a22_inverse(x, ...)

.get_tilde_a21(x)
}
\arguments{
\item{x}{a fitted \code{DirectAdjusted} model}

\item{...}{
  Arguments passed on to \code{\link[=.get_a22_inverse]{.get_a22_inverse}}, \code{\link[=.get_a21]{.get_a21}}
  \describe{
    \item{\code{}}{}
  }}
}
\value{
\code{.get_a22_inverse()}/\code{.get_tilde_a22_inverse()}: A
\eqn{k\times k} matrix where k denotes the number of parameters in the ITT
effect model

\code{.get_a11_inverse()}: A \eqn{p\times p} matrix where the
\eqn{p} is the dimension of the covariance adjustment model including an
intercept

\code{.get_a21()}/\code{.get_tilde_a21()}: A \eqn{k\times p} matrix
where the number of rows are given by the dimension of the ITT effect
estimating equations and the number of columns are given by the number of
terms in the covariance adjustment model
}
\description{
(Internal) Compute variance blocks
}
\details{
\eqn{A_{22}^{-1}} is the inverse observed Fisher information of the
ITT effect estimating equations scaled by \eqn{n_{\mathcal{Q}}}.

\eqn{A_{11}^{-1}} is the inverse of the gradient of the covariance
adjustment model estimating equations scaled by \eqn{n_{\mathcal{C}}^{-1}}.

\eqn{A_{21}} is the gradient of the ITT effect estimating equations
scaled by \eqn{n_{\mathcal{Q}}^{-1}} taken with respect to the covariance
adjustment model parameters. This matrix is the crossproduct of the
prediction gradient for the units of observation in \eqn{\mathcal{Q}} and
the model matrix of the ITT effect estimating eqations.

\eqn{\tilde{A}_{22}^{-1}} is the inverse observed Fisher
information of the ITT effect estimating equations scaled by \eqn{n}. This
function wraps around the function \code{.get_a22_inverse()} that produces
\eqn{A_{22}^{-1}}, where \eqn{A_{22}=\frac{n}{n_{\mathcal{Q}}}\tilde{A}_{22}}.

\eqn{\tilde{A}_{21}} is the gradient of the ITT effect estimating
equations scaled by \eqn{n^{-1}} taken with respect to the covariance
adjustment model parameters. This function wraps around \code{.get_a21()},
which produces \eqn{A_{21}}, where \eqn{A_{21} = \frac{n_{\mathcal{Q}}}{n}
  \tilde{A}_{21}}.
}
\keyword{internal}
