% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SandwichLayer.R
\docType{class}
\name{SandwichLayer-class}
\alias{SandwichLayer-class}
\title{(Internal) model predictions with more model artifacts, as S4 object}
\description{
Contains PreSandwichLayer class.  Only additional slots listed here.
}
\section{Slots}{

\describe{
\item{\code{keys}}{a data.frame}

\item{\code{StudySpecification}}{a StudySpecification}
}}

\keyword{internal}
