% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CovAdjPrediction.R
\name{show,CovAdjPrediction-method}
\alias{show,CovAdjPrediction-method}
\title{Show a CovAdjPrediction}
\usage{
\S4method{show}{CovAdjPrediction}(object)
}
\arguments{
\item{object}{CovAdjPredictionDesign object}
}
\value{
an invisible copy of \code{object}
}
\description{
Show a CovAdjPrediction
}
