% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design.R
\name{show,Design-method}
\alias{show,Design-method}
\title{Show a \code{Design}}
\usage{
\S4method{show}{Design}(object)
}
\arguments{
\item{object}{\code{Design} object, usually a result of a call to
\code{\link[=rct_design]{rct_design()}}, \code{\link[=obs_design]{obs_design()}}, or \code{\link[=rd_design]{rd_design()}}.}
}
\value{
\code{object}, invisibly.
}
\description{
Display information about a \code{Design} object
}
