% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DesignStructure.R
\name{show,DesignStructure-method}
\alias{show,DesignStructure-method}
\title{Show a \code{DesignStructure}}
\usage{
\S4method{show}{DesignStructure}(object)
}
\arguments{
\item{object}{\code{DesignStructure} object}
}
\value{
an invisible copy of \code{object}
}
\description{
Show a \code{DesignStructure}
}
