% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DirectAdjusted.R
\name{show,DirectAdjusted-method}
\alias{show,DirectAdjusted-method}
\title{Show a \code{DirectAdjusted}}
\usage{
\S4method{show}{DirectAdjusted}(object)
}
\arguments{
\item{object}{\code{DirectAdjusted} object, usually a result of a call to
\code{\link[=lmitt]{lmitt()}}.}
}
\value{
\code{object}, invisibly.
}
\description{
Display information about a \code{DirectAdjusted} object
}
