% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbph.R
\name{show,DirectAdjustedPBPH-method}
\alias{show,DirectAdjustedPBPH-method}
\title{Show an DirectAdjustedPBPH}
\usage{
\S4method{show}{DirectAdjustedPBPH}(object)
}
\arguments{
\item{object}{DirectAdjustedPBPH object}
}
\value{
an invisible copy of \code{object}
}
\description{
Show an DirectAdjustedPBPH
}
