% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lmitted.R
\name{show,Lmitted-method}
\alias{show,Lmitted-method}
\title{Show an Lmitted}
\usage{
\S4method{show}{Lmitted}(object)
}
\arguments{
\item{object}{Lmitted object}
}
\value{
an invisible copy of \code{object}
}
\description{
Show an Lmitted
}
