% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SandwichLayer.R
\name{show,SandwichLayer-method}
\alias{show,SandwichLayer-method}
\title{Show a SandwichLayer}
\usage{
\S4method{show}{SandwichLayer}(object)
}
\arguments{
\item{object}{SandwichLayer object}
}
\value{
an invisible copy of \code{object}
}
\description{
Show a SandwichLayer
}
