% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StudySpecification.R
\name{show,StudySpecification-method}
\alias{show,StudySpecification-method}
\title{Show a \code{StudySpecification}}
\usage{
\S4method{show}{StudySpecification}(object)
}
\arguments{
\item{object}{\code{StudySpecification} object, usually a result of a call
to \code{\link[=rct_spec]{rct_spec()}}, \code{\link[=obs_spec]{obs_spec()}}, or \code{\link[=rd_spec]{rd_spec()}}.}
}
\value{
\code{object}, invisibly.
}
\description{
Display information about a \code{StudySpecification} object
}
\examples{
library(sandwich)
library(propertee)
data(schooldata)
data(studentdata)
spec <- rct_spec(treatment ~ unitid(schoolid), schooldata)
show(spec)
summary(spec)
tm <- lmitt(math ~ 1, spec, studentdata, weights = "ate")
show(tm)
summary(tm)
bread(tm)
ef <- estfun(tm)
head(ef)
covadj_mod <- lm(math ~ gpa, studentdata)
tm <- lmitt(math ~ 1, spec, studentdata, weights = "ate",
            offset = cov_adj(covadj_mod))
bread(tm)
ef <- estfun(tm)
head(ef)
vcov(tm)
vcov(tm, type = "HC1")
vcov(tm, cov_adj_rcorrect = "HC1")
confint(tm, 2)
}
