% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teeMod.R
\name{show,teeMod-method}
\alias{show,teeMod-method}
\title{Show a \code{teeMod}}
\usage{
\S4method{show}{teeMod}(object)
}
\arguments{
\item{object}{\code{teeMod} object, usually a result of a call to
\code{\link[=lmitt]{lmitt()}}.}
}
\value{
\code{object}, invisibly.
}
\description{
Display information about a \code{teeMod} object
}
\examples{
data(schooldata)
data(studentdata)
spec <- rct_spec(treatment ~ unitid(schoolid), schooldata)
tm <- lmitt(math ~ 1, spec, studentdata, weights = "ate")
show(tm)
summary(tm)

# with covariance adjustment
covadj_mod <- lm(math ~ gpa, studentdata)
tm <- lmitt(math ~ 1, spec, studentdata, weights = "ate",
            offset = cov_adj(covadj_mod))
show(tm)
summary(tm)
}
