% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simdata}
\alias{simdata}
\title{Simulated data with unit of assignment level treatment assignment}
\format{
A data.frame with 100 rows and 7 columns.
\itemize{
\item uoa1 First level unit of assignment ID
\item uoa2 Second level unit of assignment ID
\item bid Block ID
\item force Forcing variable
\item z Binary treatment indicator
\item o 4-level ordered treatment variable
\item dose Dose treatment variable
\item x Some predictor
\item y Some outcome
}
}
\usage{
simdata
}
\description{
Simulated data with unit of assignment level treatment assignment
}
\keyword{dataset}
