% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{STARdata}
\alias{STARdata}
\title{STAR data}
\format{
A \code{data.frame} with 11,598 rows and 47 columns.
}
\usage{
STARdata
}
\description{
Tennessee’s Student-Teacher Achievement Ratio (STAR) data set.
This is a copy of the data \code{AER::STAR}.
}
\keyword{dataset}
