% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StudySpecificationStructure.R
\name{get_structure}
\alias{get_structure}
\alias{show,StudySpecificationStructure-method}
\title{\code{StudySpecification} Structure Information}
\usage{
get_structure(specification)

\S4method{show}{StudySpecificationStructure}(object)
}
\arguments{
\item{specification}{a \code{StudySpecification} object}

\item{object}{a \code{StudySpecificationStructure} object, typically the
output of \code{get_structure}}
}
\value{
A \code{StudySpecificationStructure} object containing the structure
of the \code{specification} as a \code{data.frame}.
}
\description{
Obtaining a \code{data.frame} which encodes the specification
information.
}
\examples{
data(simdata)
spec <- rct_spec(z ~ uoa(uoa1, uoa2) + block(bid), data = simdata)
get_structure(spec)
}
