% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DirectAdjusted.R
\name{summary,DirectAdjusted-method}
\alias{summary,DirectAdjusted-method}
\title{Summary of \code{DirectAdjusted} object}
\usage{
\S4method{summary}{DirectAdjusted}(object, ...)
}
\arguments{
\item{object}{DirectAdjusted}

\item{...}{Additional arguments to \code{vcovDA()}, such as the desired
finite sample heteroskedasticity-robust standard error adjustment.}
}
\value{
summary object based from \code{stats:::summary.lm()}
}
\description{
Summary of \code{DirectAdjusted} object
}
\details{
If a \code{DirectAdjusted} object is fit with a \code{SandwichLayer}
offset, then the usual \code{stats::summary.lm()} output is enhanced by
the use of covariance-adjusted sandwich standard errors, with t-test values
recalculated to reflect the new standard errors.
}
