% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.DirectAdjusted.R
\name{summary.DirectAdjusted}
\alias{summary.DirectAdjusted}
\title{Create summary of \code{DirectAdjusted} object}
\usage{
\method{summary}{DirectAdjusted}(object, ...)
}
\arguments{
\item{object}{DirectAdjusted object}

\item{...}{Other args}
}
\value{
object of class \code{summary.DirectAdjusted}
}
\description{
Create summary of \code{DirectAdjusted} object
}
