% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terms.lmitt.formula.R
\name{terms.lmitt.formula}
\alias{terms.lmitt.formula}
\title{Build a terms object from a \code{lmitt.formula}}
\usage{
\method{terms}{lmitt.formula}(x, specification, ...)
}
\arguments{
\item{x}{\code{lmitt.formula}.}

\item{specification}{\code{StudySpecification}.}

\item{...}{Additional arguments to inform attributes stored in the
\code{lmitt.terms} object.}
}
\value{
A \code{lmitt.terms} object.
}
\description{
This method parses the terms in a \code{lmitt.formula} and creates a
terms-like object that is used to generate the model frame for an
\code{lmitt()} fit.
}
\details{
The formula associated with the output--what one obtains calling
\code{formula()} on the object--is the \code{lmitt.formula} passed to
\code{x} (see \code{formula.lmitt.terms()}), but the attributes of the output
reflect the formula stored in the \code{model_form} attribute. This allows
\code{model.frame()} to build a model frame based on \code{model_form}, which
has the exact terms to be included in the model fit, while future
\code{terms.lmitt.terms()} calls have the information from the
\code{lmitt.formula} to build the model frame from new data. The
\code{model_form} and \code{absorb} attributes are specific to the
\code{lmitt.terms} class, where the latter informs future
\code{model.matrix.lmitt.terms()} calls to block-center its columns. Since
\code{model.frame()} and \code{model.matrix()} are S3 generics,
\code{lmitt.terms} is not a defined S4 class, but rather manually set within
\code{terms.lmitt.formula()}.
}
\examples{
data(simdata)
spec <- rct_spec(z ~ 1, simdata)
form <- build_lmitt_formula("y", "1", rhstype = "intercept", absorb = FALSE)
terms(form, spec)
}
