% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terms.lmitt.formula.R
\name{terms.lmitt.terms}
\alias{terms.lmitt.terms}
\title{Return an \code{lmitt.terms} object with \code{terms()}}
\usage{
\method{terms}{lmitt.terms}(x, ...)
}
\arguments{
\item{x}{\code{lmitt.terms}.}

\item{...}{Arguments passed to \code{...} do not modify the
\code{lmitt.terms} object given.}
}
\value{
The \code{lmitt.terms} object passed to \code{x} unchanged.
}
\description{
Return an \code{lmitt.terms} object with \code{terms()}
}
\examples{
data(simdata)
spec <- rct_spec(z ~ 1, simdata)
form <- build_lmitt_formula("y", "1", rhstype = "intercept", absorb = FALSE)
tt <- terms(form, spec)
stats::terms(tt)
}
