% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DirectAdjusted.R
\name{treatment,DirectAdjusted-method}
\alias{treatment,DirectAdjusted-method}
\title{Identify treatment variable in DirectAdjusted object}
\usage{
\S4method{treatment}{DirectAdjusted}(x, ...)
}
\arguments{
\item{x}{DirectAdjusted model}

\item{...}{Ignored}
}
\value{
Name of treatment in model.
}
\description{
Identify treatment variable in DirectAdjusted object
}
\examples{
data(simdata)
des <- rct_design(z ~ unitid(cid1, cid2), data = simdata)
mod <- lm(y ~ z, data = simdata, weights = ett(des))
damod <- as.DirectAdjusted(mod)
damod$coef[treatment(damod)]
des2 <- rct_design(dose ~ unitid(cid1, cid2), data = simdata,
                   dichotomy = dose > 200 ~ . )
mod2 <- lm(y ~ adopters(), data = simdata, weights = ett(des2))
damod2 <- as.DirectAdjusted(mod2)
damod2$coef[treatment(damod2)]
}
