% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lmitted.R
\name{treatment_name}
\alias{treatment_name}
\title{Identify treatment variable in \code{Lmitted} object}
\usage{
treatment_name(x)
}
\arguments{
\item{x}{\code{Lmitted} model}
}
\value{
Name of treatment in model.
}
\description{
Identify treatment variable in \code{Lmitted} object
}
\examples{
data(simdata)
des <- rct_design(z ~ unitid(cid1, cid2), data = simdata)
mod <- lm(y ~ z, data = simdata, weights = ett(des))
damod <- as.lmitt(mod)
damod$coef[treatment_name(damod)]
des2 <- rct_design(dose ~ unitid(cid1, cid2), data = simdata,
                   dichotomy = dose > 200 ~ . )
mod2 <- lm(y ~ adopters(), data = simdata, weights = ett(des2))
damod2 <- as.lmitt(mod2)
damod2$coef[treatment_name(damod2)]
}
