% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design.R
\name{treatment_table}
\alias{treatment_table}
\title{treatment group table}
\usage{
treatment_table(design, ...)
}
\arguments{
\item{design}{A Design object}

\item{...}{add'l optional arguments to \code{table}}
}
\value{
a table of treatment by units
}
\description{
Returns a table of number of units of assignment in each treatment group,
sorted by the size of the groups
}
