% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design.R
\name{treatmentTable}
\alias{treatmentTable}
\title{treatment group table}
\usage{
treatmentTable(design, ...)
}
\arguments{
\item{design}{A Design object}

\item{...}{add'l optional arguments to \code{table}}
}
\value{
a table of class \code{treatmentTable}
}
\description{
Returns a table of number of units of assignment in each treatment group,
sorted by the size of the groups
}
