% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SandwichLayerVariance.R
\name{vcovDA}
\alias{vcovDA}
\alias{.vcov_CR0}
\alias{.vcov_HC0}
\alias{.vcov_MB0}
\alias{.vcov_CR1}
\alias{.vcov_HC1}
\alias{.vcov_MB1}
\title{Compute robust sandwich variance estimates with optional covariance adjustment}
\usage{
vcovDA(x, type = "CR0", cluster = NULL, ...)

.vcov_CR0(x, ...)

.vcov_HC0(x, ...)

.vcov_MB0(x, ...)

.vcov_CR1(x, ...)

.vcov_HC1(x, ...)

.vcov_MB1(x, ...)
}
\arguments{
\item{x}{a fitted \code{DirectAdjusted} model}

\item{type}{a string indicating the desired variance estimator}

\item{cluster}{a string or character vector of column names indicating columns
to cluster standard errors by. With prior covariance adjustment, columns
must appear in both the covariance adjustment and direct adjustment samples.
Default is NULL, which uses unit of assignment columns in the \code{Design}
slot of the \code{DirectAdjusted} model.}

\item{...}{arguments to be passed to the internal variance estimation
function.}
}
\value{
A \eqn{2\times 2} matrix corresponding to an intercept and the
treatment variable in the direct adjustment model
}
\description{
Compute robust sandwich variance estimates with optional covariance adjustment
}
\details{
Supported \code{type} include:
\itemize{
\item \code{"MB0"}, \code{"HC0"}, and \code{"CR0"} for model-based HC0 standard errors
\item \code{"MB1"}, \code{"HC1"}, and \code{"CR1"} for model-based standard errors
with HC1 corrections based on the direct adjustment estimate i.e.,
\eqn{n/(n - 2)} for \code{"MB1"} and \code{"HC1"}, and for \code{"CR1"},
\eqn{g\cdot(n-1)/((g-1)\cdot(n-2))}, where \eqn{g} is the number of clusters
in the direct adjustment sample.
}

To create your own \code{type}, simply define a function \code{.vcov_XXX}.
\code{type = "XXX"} will now use your method. Your method should return a
matrix of appropriate dimension, with \code{attribute} \code{type = "XXX"}.
}
\keyword{internal}
