% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design.R
\name{var_names}
\alias{var_names}
\title{Extract names of Design variables}
\usage{
var_names(x, type)
}
\arguments{
\item{x}{Design x}

\item{type}{one of "t", "u", "b", "f"; for "treatment",
"unit_of_assignment", "block", and "forcing"}
}
\value{
character vector of variable names of the given type
}
\description{
Extract names of Design variables
}
