% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design.R
\name{var_table}
\alias{var_table}
\title{variable identification table}
\usage{
var_table(design, ..., compress = TRUE, report_all = FALSE)
}
\arguments{
\item{design}{A Design object}

\item{...}{add'l optional arguments to \code{table}}

\item{compress}{Should multiple variables be compressed into a
comma-separated string? Default TRUE.}

\item{report_all}{Should we report all possible structures even if they
don't exist in the Design? Default FALSE.}
}
\value{
a table of variables in the Design structure
}
\description{
Returns a table containing the variables identified in each structure
}
