% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design.R
\name{varNames}
\alias{varNames}
\title{Extract names of Design variables}
\usage{
varNames(x, type)
}
\arguments{
\item{x}{Design x}

\item{type}{one of "t", "c", "b", "f"; for "treatment", "cluster", "block",
and "forcing"}
}
\value{
character vector of variable names of the given type
}
\description{
Extract names of Design variables
}
