% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design.R
\name{varTable}
\alias{varTable}
\title{variable identification table}
\usage{
varTable(design, ..., compress = TRUE, reportAll = FALSE)
}
\arguments{
\item{design}{A Design object}

\item{...}{add'l optional arguments to \code{table}}

\item{compress}{Should multiple variables be compressed into a
comma-separated string? Default TRUE.}

\item{reportAll}{Should we report all possible structures even if they don't
exist in the Design? Default FALSE.}
}
\value{
a table of variables in the Design structure
}
\description{
Returns a table containing the variables identified in each structure
}
