% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DirectAdjusted.R
\name{vcov,DirectAdjusted-method}
\alias{vcov,DirectAdjusted-method}
\title{Variance-Covariance matrix}
\usage{
\S4method{vcov}{DirectAdjusted}(object, ...)
}
\arguments{
\item{object}{DirectAdjusted}

\item{...}{Add'l arguments}
}
\value{
Variance-Covariance matrix
}
\description{
Variance-Covariance matrix
}
