% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lmitted.R
\name{vcov,Lmitted-method}
\alias{vcov,Lmitted-method}
\title{Variance-Covariance matrix}
\usage{
\S4method{vcov}{Lmitted}(object, ...)
}
\arguments{
\item{object}{Lmitted}

\item{...}{Add'l arguments}
}
\value{
Variance-Covariance matrix
}
\description{
Variance-Covariance matrix
}
