% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SandwichLayerVariance.R
\name{vcovDA}
\alias{vcovDA}
\title{Compute direct adjustment sandwich variance estimator}
\usage{
vcovDA(x, ...)
}
\arguments{
\item{x}{A \code{Lmitted} model}

\item{...}{Arguments to be passed to sandwich::meatCL}
}
\value{
\code{vcovDA()}: A \eqn{2\times 2} matrix where the dimensions are
given by the intercept and treatment variable terms in the direct adjustment
model
}
\description{
Compute direct adjustment sandwich variance estimator
}
\details{
\code{vcovDA()} provides covariance-adjusted cluster-robust variance
estimates for treatment effects.
}
