% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeightedDesign.R
\name{subset,WeightedDesign-method}
\alias{subset,WeightedDesign-method}
\alias{[,WeightedDesign-method}
\title{\code{WeightedDesign} subsetting}
\usage{
\S4method{subset}{WeightedDesign}(x, subset)

\S4method{[}{WeightedDesign}(x, i)
}
\arguments{
\item{x}{\code{WeightedDesign} object}

\item{subset}{Logical vector identifying values to keep or drop}

\item{i}{indices specifying elements to extract or replace. See
\code{help("[")} for further details.}
}
\value{
\code{x} subset by \code{i}

\code{x} subset by \code{i}
}
\description{
\code{WeightedDesign} subsetting
}
