% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeightedStudySpecification.R
\docType{class}
\name{WeightedStudySpecification-class}
\alias{WeightedStudySpecification-class}
\title{(Internal) Modeling weights with an accompanying StudySpecification}
\description{
(Internal) Modeling weights with an accompanying StudySpecification
}
\details{
\verb{@target} is used for calculation purpose; defining what weight to calculate
\verb{@weightAlias} is only to store the alias used in creation of the
weights in case we want to report it later.
}
\section{Slots}{

\describe{
\item{\code{.Data}}{numeric vector of modeling weights}

\item{\code{StudySpecification}}{a StudySpecification}

\item{\code{target}}{character string, e.g. "ate"}

\item{\code{weightAlias}}{alias for target appearing in an originating call}

\item{\code{dichotomy}}{formula describing a treatment/comparison dichotomy}
}}

\keyword{internal}
