% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeightedStudySpecification.R
\name{subset,WeightedStudySpecification-method}
\alias{subset,WeightedStudySpecification-method}
\alias{[,WeightedStudySpecification,ANY,ANY,ANY-method}
\title{\code{WeightedStudySpecification} subsetting}
\usage{
\S4method{subset}{WeightedStudySpecification}(x, subset)

\S4method{[}{WeightedStudySpecification,ANY,ANY,ANY}(x, i)
}
\arguments{
\item{x}{\code{WeightedStudySpecification} object}

\item{subset}{Logical vector identifying values to keep or drop}

\item{i}{indices specifying elements to extract or replace. See
\code{help("[")} for further details.}
}
\value{
A \code{WeightedStudySpecification} object which is a subsetted
version of \code{x}.
}
\description{
Provides functionality to subset the weights of a
\code{WeightedStudySpecification} object.
}
\examples{
data(schooldata)
data(studentdata)
spec <- rct_spec(treatment ~ unitid(schoolid) + block(state), schooldata)
wt <- ate(spec, data = schooldata)
wt.subset <- wt[seq_len(4)]
identical(subset(wt, c(rep(TRUE, 4), rep(FALSE, nrow(schooldata)-4))),
          wt.subset)
show(wt.subset)
weights(wt.subset)
wt2 <- ett(spec, data = schooldata)
wt2.subset <- wt2[seq_len(4)]
can_add <- tryCatch(wt.subset + wt2.subset, error = function(e) FALSE)
can_add
can_multiply <- tryCatch(wt.subset * wt2.subset, error = function(e) FALSE)
can_multiply
can_divide <- tryCatch(wt.subset / wt2.subset, error = function(e) FALSE)
can_divide
other.wt <- c(1, 1, 0, 1)
can_multiply2 <- tryCatch(wt.subset * other.wt, error = function(e) FALSE)
can_multiply2
can_divide2 <- tryCatch(other.wt / wt.subset, error = function(e) FALSE)
can_divide2
}
