% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeightedStudySpecification.R
\name{weights,WeightedStudySpecification-method}
\alias{weights,WeightedStudySpecification-method}
\title{Extract Weights from \code{WeightedStudySpecification}}
\usage{
\S4method{weights}{WeightedStudySpecification}(object, ...)
}
\arguments{
\item{object}{a \code{WeightedStudySpecification} object}

\item{...}{Ignored}
}
\value{
A numeric \code{vector} of the weights
}
\description{
A \code{WeightedStudySpecification} object contains a numeric
vector with a few additional slots, this extracts only the numeric vector.
}
\examples{
data(schooldata)
data(studentdata)
spec <- rct_spec(treatment ~ unitid(schoolid) + block(state), schooldata)
wt <- ate(spec, data = schooldata)
wt.subset <- wt[seq_len(4)]
identical(subset(wt, c(rep(TRUE, 4), rep(FALSE, nrow(schooldata)-4))),
          wt.subset)
show(wt.subset)
weights(wt.subset)
wt2 <- ett(spec, data = schooldata)
wt2.subset <- wt2[seq_len(4)]
can_add <- tryCatch(wt.subset + wt2.subset, error = function(e) FALSE)
can_add
can_multiply <- tryCatch(wt.subset * wt2.subset, error = function(e) FALSE)
can_multiply
can_divide <- tryCatch(wt.subset / wt2.subset, error = function(e) FALSE)
can_divide
other.wt <- c(1, 1, 0, 1)
can_multiply2 <- tryCatch(wt.subset * other.wt, error = function(e) FALSE)
can_multiply2
can_divide2 <- tryCatch(other.wt / wt.subset, error = function(e) FALSE)
can_divide2
}
