% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SandwichLayer.R
\name{as.SandwichLayer}
\alias{as.SandwichLayer}
\title{Convert a PreSandwichLayer to a SandwichLayer via a Design Object}
\usage{
as.SandwichLayer(x, design, by = NULL, Q_data = NULL)
}
\arguments{
\item{x}{a \code{PreSandwichLayer} object.}

\item{design}{a \code{Design} object created by one of \code{rct_design()},
\code{rd_design()}, or \code{obs_design()}.}

\item{by}{character vector or list; optional. Specifies column names that
appear in both the covariance adjustment and quasiexperimental dataframes
and can be used to match observations in case the two overlap. Names
represent variables in the Design; values represent variables in the data.
Defaults to NULL, in which case unit of assignment columns indicated in
the Design will be used to match observations in the samples.}

\item{Q_data}{optional; quasiexperimental dataframe to allow for merging
to the covariance adjustment data via the \code{by} argument. If \code{by} is not NULL
and \code{Q_data} is NULL, the function will search through the call stack to find
the quasiexperimental data; if the search is unsuccessful, the function will
use the \code{design} data to attempt to merge.}
}
\value{
a \code{SandwichLayer} object
}
\description{
Convert a PreSandwichLayer to a SandwichLayer via a Design Object
}
