% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SandwichLayer.R
\name{as.SandwichLayer}
\alias{as.SandwichLayer}
\title{Convert a PreSandwichLayer to a SandwichLayer via a Design Object}
\usage{
as.SandwichLayer(x, design, by = NULL, envir = parent.frame())
}
\arguments{
\item{x}{a \code{PreSandwichLayer} object.}

\item{design}{a \code{Design} object created by one of \code{rct_design()},
\code{rd_design()}, or \code{obs_design()}.}

\item{by}{optional; named vector or list connecting names of cluster/unit of
assignment variables in \code{design} to cluster/unit of assignment
variables in \code{data}. Names represent variables in the Design; values
represent variables in the data. Only needed if variable names differ.}

\item{envir}{an \code{Environment} object. The environment in which to find
the covariance model data to instantiate. By default, \code{parent.frame()}.}
}
\value{
a \code{SandwichLayer} object
}
\description{
Convert a PreSandwichLayer to a SandwichLayer via a Design Object
}
