% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.lmitt.R
\name{as.lmitt}
\alias{as.lmitt}
\alias{as.DirectAdjusted}
\title{Convert \code{lm} object into \code{DirectAdjusted}}
\usage{
as.lmitt(x, design = NULL)

as.DirectAdjusted(x, design = NULL)
}
\arguments{
\item{x}{\code{lm} object with weights containing a \code{WeightedDesign}}

\item{design}{Optional, explicitly specify the \code{Design} to be used. If
the \code{Design} is specified elsewhere in \code{x} (e.g. passed as an
argument to any of \code{ate()}, \code{ett()}, \code{cov_adj()} or
\code{assigned()}) it will be found automatically and does not need to be
passed here as well. (If different \code{Design} objects are passed
(either through the \code{lm} in weights or covariance adjustment, or
through this argument), an error will be produced.)}
}
\value{
\code{DirectAdjusted} object
}
\description{
Convert \code{lm} object into \code{DirectAdjusted}
}
