% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assigned.R
\name{assigned}
\alias{assigned}
\alias{adopters}
\alias{a.}
\alias{z.}
\title{Obtain Treatment from Design}
\usage{
assigned(design = NULL, data = NULL)

adopters(design = NULL, data = NULL)

a.(design = NULL, data = NULL)

z.(design = NULL, data = NULL)
}
\arguments{
\item{design}{Optional \code{Design}. If the \code{Design} can't be
identified in the model (usually because neither weights (\code{ate()} or
\code{ett()}) nor a covariate adjustment model (\code{cov_adj()}) are
found), the \code{Design} can be passed diretly.}

\item{data}{Optional data set. By default \code{assigned()} will attempt to
identify the appropriate data, if this fails (or you want to overwrite
it), you can pass the data here.}
}
\value{
The treatment variable to be placed in the regression formula.
}
\description{
When included as a predictor in a model formula where weights (via
\code{ate()} or \code{ett()} being passed to the \code{weights=} argument)
or a covariance adjustment model (via \code{cov_adj()} being passed to the
\code{offset=} argument) are included, this will extract the treatment
variable from the \code{Design}.
}
\details{
While \code{assigned()} can be used in any situation, it is most useful for
scenarios where the treatment variable is non-binary and the \code{Design}
contains a \code{Dichotomy}. For example, say \code{q} is a 3-level ordinal
treatment variable, and the binary comparison of interest is captured in
\code{dichotomy = q == 3 ~ q < 3}. If you were to fit a model including
\code{q} as a predictor, e.g. \code{lm(y ~ q, ...)}, \code{lm} would treat
\code{q} as the full ordinal variable. On the other hand, by calling
\code{lm(y ~ assigned(), weights = ate(des), ...)}, \code{assigned()} will
generate the appropriate binary variable to allow estimation of treatment
effects.

If called outside of a model call and without a \code{data} argument, this
will extract the treatment from the \code{design}.
}
