% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DirectAdjusted.R
\name{bread.DirectAdjusted}
\alias{bread.DirectAdjusted}
\title{Extract bread matrix from a \code{DirectAdjusted} model fit}
\usage{
\method{bread}{DirectAdjusted}(x)
}
\arguments{
\item{x}{DirectAdjusted object}
}
\value{
A \eqn{k\times k} matrix where k denotes the number of treatment
levels
}
\description{
Extract bread matrix from a \code{DirectAdjusted} model fit
}
\details{
Returns the expected Fisher information for the ITT effect model
fit
}
