% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DirectAdjusted.R
\name{bread.DirectAdjusted}
\alias{bread.DirectAdjusted}
\title{Extract bread matrix from a \code{DirectAdjusted} model fit}
\usage{
\method{bread}{DirectAdjusted}(x, ...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=.get_a22_inverse]{.get_a22_inverse}}
  \describe{
    \item{\code{x}}{a fitted \code{DirectAdjusted} model}
  }}
}
\value{
\code{.get_a22_inverse()}/\code{.get_tilde_a22_inverse()}: A
\eqn{k\times k} matrix where k denotes the number of parameters in the ITT
effect model

\code{.get_a11_inverse()}: A \eqn{p\times p} matrix where the
\eqn{p} is the dimension of the covariance adjustment model including an
intercept

\code{.get_a21()}/\code{.get_tilde_a21()}: A \eqn{k\times p} matrix
where the number of rows are given by the dimension of the ITT effect
estimating equations and the number of columns are given by the number of
terms in the covariance adjustment model
}
\description{
Extract bread matrix from a \code{DirectAdjusted} model fit
}
\details{
This function is a thin wrapper around \code{.get_tilde_a22_inverse()}.
}
