% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DirectAdjusted.R
\name{bread.DirectAdjusted}
\alias{bread.DirectAdjusted}
\title{Extract bread matrix from a \code{DirectAdjusted} model fit}
\usage{
\method{bread}{DirectAdjusted}(x, ...)
}
\arguments{
\item{x}{a fitted \code{DirectAdjusted} object}

\item{...}{arguments passed to methods}
}
\value{
A \eqn{k\times k} matrix where k denotes the number of parameters
in the ITT effect model. This corresponds to the Hessian of the ITT effect
model estimating equations defined in our accompanying documentation.
}
\description{
Extract bread matrix from a \code{DirectAdjusted} model fit
}
