% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmitt.formula.R
\name{build_lmitt_formula}
\alias{build_lmitt_formula}
\title{Build the formula for \code{lmitt()}}
\usage{
build_lmitt_formula(moderator_var, moderator_type, absorb, dichotomy = NULL)
}
\arguments{
\item{moderator_var}{character, name of the moderator variable. May be "1"
if no moderator variable was provided in the original formula.}

\item{moderator_type}{character, one of \code{"intercept"},
\code{"continuous"}, or \code{"categorical"}.}

\item{absorb}{logical, \code{TRUE} if moderator variable is to be block-
centered.}

\item{dichotomy}{formula, specifies the dichotomization of a non-binary
treatment assignment variable. Defaults to NULL if the treatment assignment
variable is already binary.}
}
\value{
Returns a formula that has a term for the treatment assignment
variable and, if provided in \code{lmitt()}, a moderator term that will be
appropriately block-centered (as indicated by the \code{absorb} argument of
\code{lmitt()}) and interacted with the treatment assignment variable (based
on whether the moderator variable is continuous or categorical)
}
\description{
Build the formula for \code{lmitt()}
}
