% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CombinedWeightedDesign.R
\name{c,WeightedDesign-method}
\alias{c,WeightedDesign-method}
\title{Concatenate weights}
\usage{
\S4method{c}{WeightedDesign}(x, ..., force_dichotomy_equal = FALSE)
}
\arguments{
\item{...}{any number of \code{WeightedDesign} objects with equivalent
\code{Design}.}

\item{force_dichotomy_equal}{if \code{FALSE} (default), \code{Design}s are
considered equivalent even if their \code{dichotomy} differs. If
\code{TRUE}, \code{@dichotomy} must also be equal.}
}
\description{
Given several variations of weights generated from a single \code{Design},
combine into a single weight.
}
\details{
Concatenating \code{WeightedDesign}s with \code{c()} requires both
individual \code{WeightedDesign}s to come from the same \code{Design}
(except \code{dichotomy}, see below) and have the target (e.g all created
with \code{ate()} or all created with \code{ett()}, no mixing-and-matching).
All arguments to \code{c()} must be \code{WeightedDesign}.

One exception is when concatenting \code{WeightedDesign}s with the same
\code{Design} but different dichotomies. There may be cases where the
treatment is continuous or has multiple levels, and there is a need to
combine the weights from the same general design, but with different
dichotomys. Therefore multiple \code{WeightedDesign}s can be combined if
they are identical except for their \code{@dichotomy} slots. The resulting
object will be a \code{CombinedWeightedDesign} which tracks all individual
\code{dichotomy}.
}
