% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SandwichLayerVariance.R
\name{cluster_iss}
\alias{cluster_iss}
\title{Use a rank-1 update to cheaply compute inverse symmetric square roots of
projection matrices}
\usage{
cluster_iss(
  tm,
  cluster_unit,
  cluster_ids = NULL,
  cluster_var = NULL,
  exclude = na.action(tm),
  ...
)
}
\arguments{
\item{exclude}{index of units to exclude from computing the correction; for
example, if they're NA's}
}
\description{
Use a rank-1 update to cheaply compute inverse symmetric square roots of
projection matrices
}
\keyword{internal}
