% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DirectAdjusted.R
\name{confint.DirectAdjusted}
\alias{confint.DirectAdjusted}
\title{Confidence intervals with standard errors provided by \code{vcov.DirectAdjusted()}}
\usage{
\method{confint}{DirectAdjusted}(object, parm, level = 0.95, ...)
}
\arguments{
\item{object}{a fitted \code{DirectAdjusted} model}

\item{parm}{a specification of which parameters are to be given confidence
intervals, either a vector of numbers or a vector of names. If missing,
all parameters are considered.}

\item{level}{the confidence level required.}

\item{...}{additional arguments to pass to \code{vcov.DirectAdjusted()}}
}
\value{
A matrix (or vector) with columns giving lower and upper confidence
limits for each parameter. These will be labelled as (1-level)/2 and 1 -
(1-level)/2 in \% (by default 2.5\% and 97.5\%)
}
\description{
An S3method for \code{stats::confint} that uses standard errors computed
using \code{vcov.DirectAdjusted()}. Additional arguments passed to this
function, such as \code{cluster} and \code{type}, specify the arguments of
the \code{vcov.DirectAdjusted()} call.
}
