% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_adj.R
\name{cov_adj}
\alias{cov_adj}
\title{Covariance Adjustment for Treatment Estimation}
\usage{
cov_adj(model)
}
\arguments{
\item{model}{Any model which supports a \code{predict} function}
}
\value{
Covariate adjusted outcomes
}
\description{
Covariance Adjustment for Treatment Estimation
}
