% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design.R
\name{treatment}
\alias{treatment}
\alias{treatment,Design-method}
\alias{treatment<-}
\alias{treatment<-,Design-method}
\alias{clusters}
\alias{clusters,Design-method}
\alias{clusters<-}
\alias{clusters<-,Design-method}
\alias{blocks}
\alias{blocks,Design-method}
\alias{blocks<-}
\alias{blocks<-,Design-method}
\alias{forcings}
\alias{forcings,Design-method}
\alias{forcings<-}
\alias{forcings<-,Design-method}
\title{Extract and Replace elements of Design}
\usage{
treatment(x)

\S4method{treatment}{Design}(x)

treatment(x) <- value

\S4method{treatment}{Design}(x) <- value

clusters(x)

\S4method{clusters}{Design}(x)

clusters(x) <- value

\S4method{clusters}{Design}(x) <- value

blocks(x)

\S4method{blocks}{Design}(x)

blocks(x) <- value

\S4method{blocks}{Design}(x) <- value

forcings(x)

\S4method{forcings}{Design}(x)

forcings(x) <- value

\S4method{forcings}{Design}(x) <- value
}
\arguments{
\item{x}{Design object}

\item{value}{Replacement. Either a vector/matrix of appropriate dimension,
or a named data.frame if renaming variable as well.}
}
\value{
data.frame containing cluster-level information
}
\description{
Extract and Replace elements of Design
}
