% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DesignAccessors.R
\name{treatment}
\alias{treatment}
\alias{treatment,Design-method}
\alias{treatment<-}
\alias{treatment<-,Design-method}
\alias{units_of_assignment}
\alias{units_of_assignment,Design-method}
\alias{units_of_assignment<-}
\alias{units_of_assignment<-,Design-method}
\alias{clusters}
\alias{clusters,Design-method}
\alias{clusters<-}
\alias{clusters<-,Design-method}
\alias{unitids}
\alias{unitids,Design-method}
\alias{unitids<-}
\alias{unitids<-,Design-method}
\alias{blocks}
\alias{blocks,Design-method}
\alias{blocks<-}
\alias{blocks<-,Design-method}
\alias{forcings}
\alias{forcings,Design-method}
\alias{forcings<-}
\alias{forcings<-,Design-method}
\title{Accessors for Design}
\usage{
treatment(x, binary = FALSE)

\S4method{treatment}{Design}(x, binary = FALSE)

treatment(x) <- value

\S4method{treatment}{Design}(x) <- value

units_of_assignment(x)

\S4method{units_of_assignment}{Design}(x)

units_of_assignment(x) <- value

\S4method{units_of_assignment}{Design}(x) <- value

clusters(x)

\S4method{clusters}{Design}(x)

clusters(x) <- value

\S4method{clusters}{Design}(x) <- value

unitids(x)

\S4method{unitids}{Design}(x)

unitids(x) <- value

\S4method{unitids}{Design}(x) <- value

blocks(x)

\S4method{blocks}{Design}(x)

blocks(x) <- value

\S4method{blocks}{Design}(x) <- value

forcings(x)

\S4method{forcings}{Design}(x)

forcings(x) <- value

\S4method{forcings}{Design}(x) <- value
}
\arguments{
\item{x}{Design object}

\item{binary}{Logical; if FALSE (default), return a \code{data.frame} containing
the named treatment variable. If TRUE and \code{x} has a formula in
\code{@dichotomy}, return a \code{data.frame} containing a binary
treatment variable with the name "z__". Has no effect if
\code{@dichotomy} is \code{NULL}.}

\item{value}{Replacement. Either a vector/matrix of appropriate dimension,
or a named data.frame if renaming variable as well.}
}
\value{
If \code{binary = FALSE} data.frame containing treatment variable.
If \code{binary = TRUE}, a vector of 0/1 indicating treatment group
membership.
}
\description{
Extract and replace elements of the Design
}
\details{
For \code{treatment()}, the value of \code{binary} also affects the
dimensions of what is return. \code{treatment(..., binary = FALSE)} (the
default) will return a \code{data.frame} with the treatment as a column, for
consistency with the other accessors. However, for \code{treatment(...,
binary = TRUE)}, the output is a vector.
}
