% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design.R
\name{rct_design}
\alias{rct_design}
\alias{rd_design}
\alias{obs_design}
\title{Specify Design}
\usage{
rct_design(formula, data, subset = NULL)

rd_design(formula, data, subset = NULL)

obs_design(formula, data, subset = NULL)
}
\arguments{
\item{formula}{defines the Design components}

\item{data}{the data set.}

\item{subset}{optionally subset the data before creating the Design object}
}
\value{
a Design object of the requested type for use in further analysis
}
\description{
Generates a Design object with the given specifications.
}
\details{
Generates a randomized control treatment Design (\code{rct_design}), or an
observational Design (\code{obs_design}), or a regression discontinuity Design
(\code{rd_design}).

The formula must include exactly one of \code{unit_of_assignment()}, \code{uoa()},
\code{cluster()}, or \code{unitid()} to identify the units of assignment (one or more
variables). If defining an rd_design, the formula must also include a
\code{forcing()} entry. The formula may optionally include a \code{block()} entry as
well.
}
