% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design.R
\name{RCT_Design}
\alias{RCT_Design}
\alias{RD_Design}
\alias{Obs_Design}
\title{Specify Design}
\usage{
RCT_Design(formula, data, subset = NULL)

RD_Design(formula, data, subset = NULL)

Obs_Design(formula, data, subset = NULL)
}
\arguments{
\item{formula}{defines the design components}

\item{data}{the data set.}

\item{subset}{optionally subset the data before creating the design object}
}
\value{
a Design object of the requested type for use in further analysis
}
\description{
Generates a Design object with the given specifications.
}
\details{
Generates a randomized control treatment Design (\code{RCT_Design}), or an
observational Design (\code{Obs_Design}), or a regression discontinuity Design
(\code{RD_Design}).

The formula must include \code{cluster()} or \code{unitid()} to identify the units of
assignment (one or more variables). If defining an RD_Design, the formula
must also include a \code{forcing()} entry. The formula may optionally include a
\code{block()} entry as well.

It is strongly recommended NOT to transform variables inside the \verb{*_Design}
calls, as that can lead to mismatches between the Design objects and future
data. In other words, avoid calls such as \verb{RCT_Design(as.factor(z) ~ ...},
and instead transform the variable in all relevant data sets a prior.
}
