% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DesignStructure.R
\name{get_structure}
\alias{get_structure}
\alias{show,DesignStructure-method}
\title{Returns \code{Design} Structure Information}
\usage{
get_structure(design, binary = FALSE)

\S4method{show}{DesignStructure}(object)
}
\arguments{
\item{design}{a \code{Design} object}

\item{binary}{default \code{FALSE}. If \code{TRUE} and the design contains a
\code{dichotomy}, replace the treatment column with its binary
representation. Has no effect if \code{design} is not dichotomized.}

\item{object}{\code{DesignStructure} object}
}
\value{
A \code{data.frame} containing the structure of the \code{design}.

an invisible copy of \code{object}
}
\description{
Creates a \code{data.frame} containing the information at the unit of
assignment/unitid/cluster level.
}
