% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtable.R
\name{dtable}
\alias{dtable}
\alias{design_table}
\title{Table of elements from a \code{Design}}
\usage{
dtable(
  design,
  x,
  y = NULL,
  sort = FALSE,
  decreasing = TRUE,
  use_var_names = FALSE,
  ...
)

design_table(
  design,
  x,
  y = NULL,
  sort = FALSE,
  decreasing = TRUE,
  use_var_names = FALSE,
  ...
)
}
\arguments{
\item{design}{A \code{Design}}

\item{x}{One of "treatment", "unit of assignent", "uoa", "unitid",
"cluster", "block". Abbreviations are accepted.}

\item{y}{Optionally, another of "treatment", "unit of assignent", "uoa",
"unitid", "cluster", "block". Abbreviations are accepted. A 1-dimensional
table is produced if \code{y} is left at its default, \code{NULL}.}

\item{sort}{Ignored if \code{y} is not \code{NULL}. If \code{FALSE}
(default), one-way table is sorted according to "names" of levels. If set
to \code{TRUE}, one-way table is sorted according to values.}

\item{decreasing}{If \code{sort} is \code{TRUE}, choose whether to sort
descending (\code{TRUE}, default) or ascending (\code{FALSE}).}

\item{use_var_names}{If \code{TRUE}, name dimensions of table returned by
variable names. If \code{FALSE} (default), name by their function (e.g.
"treatment" or "blocks"). Passing the \code{dnn} argument in \code{...}
(an argument of \code{table()}) overrides whatever is requested here.}

\item{...}{other arguments to \code{table}}
}
\value{
table
}
\description{
Produces a table (1-dimensional, or 2-dimensional if \code{y}
is specified) of the elements of the \code{Design}.
}
