% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DesignUtilities.R
\name{is_dichotomized}
\alias{is_dichotomized}
\alias{has_binary_treatment}
\title{Check if Design's treatment was dichotomized}
\usage{
is_dichotomized(des)

has_binary_treatment(des)
}
\arguments{
\item{des}{A design}
}
\value{
\code{TRUE if \code{des} was dichotomized; \code{FALSE} otherwise

\code{TRUE if \code{des} has a binary treatment; \code{FALSE}
otherwise
}
\description{
A Design's treatment is considered binary if the treatment is numeric with
only values of 0/1, or \code{TRUE}/\code{FALSE}, or \code{NA}, if the Design is dichotomized.
}
