% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design.R
\name{as_RCT_Design}
\alias{as_RCT_Design}
\alias{as_Obs_Design}
\alias{as_RD_Design}
\title{Design conversion}
\usage{
as_RCT_Design(Design, ..., force = FALSE)

as_Obs_Design(Design, ..., force = FALSE)

as_RD_Design(Design, ...)
}
\arguments{
\item{Design}{Design to convert}

\item{...}{No addiitonal options at present}

\item{force}{Converting from RD to another design will error to avoid losing
the forcing variable. Setting \code{force = TRUE} allows the conversion to
automatically drop the forcing variable.}
}
\value{
The design of the updated type
}
\description{
Converst between Design types
}
