% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DesignConverters.R
\name{as_rct_design}
\alias{as_rct_design}
\alias{as_obs_design}
\alias{as_rd_design}
\title{Convert \code{Design} between types}
\usage{
as_rct_design(Design, ..., loseforcing = FALSE)

as_obs_design(Design, ..., loseforcing = FALSE)

as_rd_design(Design, data, ..., forcing)
}
\arguments{
\item{Design}{a \code{Design} to convert}

\item{...}{Ignored.}

\item{loseforcing}{converting from RD to another \code{Design} type will
error to avoid losing the forcing variable. Setting \code{loseforcing =
  TRUE} allows the conversion to automatically drop the forcing variable.
Default \code{FALSE}.}

\item{data}{converting to an RD requires adding a \code{forcing} variable,
which requires access to the original data.}

\item{forcing}{converting to an RD requires adding a \code{forcing}
variable. This should be entered as a formula which would be passed to
\code{\link[=update]{update()}}, e.g. \code{forcing = . ~ . + forcing(forcevar)}.}
}
\value{
\code{Design} of the updated type
}
\description{
Convert a \code{Design} between a observational study, a
randomized control trial, and a regression discontinuity (created from
\code{\link[=obs_design]{obs_design()}}, \code{\link[=rct_design]{rct_design()}} and \code{\link[=rd_design]{rd_design()}} respectively).
}
\examples{
des <- rct_design(z ~ unit_of_assignment(uoa1, uoa2), data = simdata)
des
as_obs_design(des)
as_rd_design(des, simdata, forcing = ~ . + forcing(force))
des2 <- rd_design(o ~ uoa(uoa1, uoa2) + forcing(force), data = simdata)
des2
# as_rct_design(des2) # this will produce an error
as_rct_design(des2, loseforcing = TRUE)
}
