% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DesignUtilities.R
\name{unit_of_assignment}
\alias{unit_of_assignment}
\alias{unitid}
\alias{cluster}
\alias{uoa}
\alias{block}
\alias{forcing}
\title{Special terms in \code{Design} creation formula}
\usage{
unit_of_assignment(...)

unitid(...)

cluster(...)

uoa(...)

block(...)

forcing(...)
}
\arguments{
\item{...}{any number of variables of the same length.}
}
\value{
the variables with appropriate labels
}
\description{
These are special function used only in the definition of Design class
objects. They identify the units of assignment, blocks and forcing
variables.
}
\details{
These functions have no use outside of the formula in creating a
\code{Design}.

\code{unit_of_assignment}, \code{uoa}, \code{cluster} and \code{unitid} are
all aliases of each other; you must include one and only one in each
\code{Design}. The choice of which to use will have no impact on any
analysis, only on some output display.
}
