% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DirectAdjusted.R
\name{.align_and_extend_estfuns}
\alias{.align_and_extend_estfuns}
\title{(Internal) Align the dimensions and rows of estimating equations matrices
from the ITT effect and covariance adjustment models}
\usage{
.align_and_extend_estfuns(x, ...)
}
\arguments{
\item{x}{a fitted \code{DirectAdjusted} object}

\item{...}{arguments passed to methods}
}
\value{
list of two matrices, one being the aligned contributions to the
estimating equations for the ITT effect model, and the other being the
aligned contributions to the covariance adjustment model
}
\description{
(Internal) Align the dimensions and rows of estimating equations matrices
from the ITT effect and covariance adjustment models
}
\keyword{internal}
