% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teeMod.R
\name{.base_S3class_estfun}
\alias{.base_S3class_estfun}
\title{(Internal) Extract empirical estimating equations from a
\code{teeMod} model using the S3 method associated with its
\code{.S3Class} slot}
\usage{
.base_S3class_estfun(x)
}
\arguments{
\item{x}{a fitted \code{teeMod} model}
}
\value{
S3 method
}
\description{
(Internal) Extract empirical estimating equations from a
\code{teeMod} model using the S3 method associated with its
\code{.S3Class} slot
}
\keyword{internal}
