% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SandwichLayerVariance.R
\name{.check_cluster_col_nas}
\alias{.check_cluster_col_nas}
\title{(Internal) Check the NA status of clustering columns in the covariance
model data}
\usage{
.check_cluster_col_nas(cluster_cols, df)
}
\arguments{
\item{cluster_cols}{vector of column names}

\item{df}{dataframe whose columns are checked for NA's}
}
\value{
A vector of column names for the valid clustering columns
}
\description{
(Internal) Check the NA status of clustering columns in the covariance
model data
}
\details{
\code{.check_cluster_col_nas} checks whether entire columns in a
dataframe contain NA values. If a column only contains NA's, the function
produces a warning indicating the column cannot be meaningfully used for
clustering.
}
\keyword{internal}
