% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SandwichLayerVariance.R
\name{.check_df_moderator_estimates}
\alias{.check_df_moderator_estimates}
\title{NaN vcovDA subgroup estimates that have insufficient degrees of freedom}
\usage{
.check_df_moderator_estimates(
  vmat,
  model,
  cluster,
  data = quote(data),
  envir = environment(formula(model))
)
}
\arguments{
\item{vmat}{variance-covariance matrix corresponding to \code{model}}

\item{model}{\code{DirectAdjusted} object}

\item{cluster}{character or factor vector providing cluster ID's for the
observations used to fit \code{model}}

\item{envir}{environment to get \code{model_data} from if it is a quote object name}

\item{model_data}{dataframe or name corresponding to the data used to fit \code{model}}
}
\value{
\code{vmat} with NaN's in the entries lacking sufficient degrees of freedom
}
\description{
NaN vcovDA subgroup estimates that have insufficient degrees of freedom
}
