% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SandwichLayerVariance.R
\name{.compute_IK_dof}
\alias{.compute_IK_dof}
\title{Compute the degrees of freedom of a sandwich standard error with HC2 correction}
\usage{
.compute_IK_dof(
  tm,
  ell,
  cluster = NULL,
  bin_y = FALSE,
  exclude = na.action(tm),
  tol = 1e-09
)
}
\description{
Compute the degrees of freedom of a sandwich standard error with HC2 correction
}
\references{
Guido W. Imbens and Michael Kolesár. "Robust Standard Errors in
Small Samples: Some Practical Advice". In: \emph{The Review of Economics and
Statistics} 98.4 (Oct. 2016), pp. 701-712.
}
\keyword{internal}
