% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SandwichLayerVariance.R
\name{.cov_mat_est}
\alias{.cov_mat_est}
\title{(Internal) Helper function for design-based meat matrix
calculation}
\usage{
.cov_mat_est(XXz, bidz)
}
\value{
estimated upper and lower bounds of covariance matrix of estimating
function vectors under either treatment or control
}
\description{
(Internal) Helper function for design-based meat matrix
calculation
}
\details{
Diagonal elements are estimated by sample variances Off-diagonal
elements are estimated using the Young's elementary inequality
}
\keyword{internal}
