% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SandwichLayerVariance.R
\name{.get_a22_inverse}
\alias{.get_a22_inverse}
\title{(Internal) Get the inverse of the A22 block of the sandwich variance estimator}
\usage{
.get_a22_inverse(x)
}
\arguments{
\item{x}{A DirectAdjusted model object}
}
\value{
A 1x1 matrix
}
\description{
(Internal) Get the inverse of the A22 block of the sandwich variance estimator
}
\details{
This block is the diagonal element of the inverse expected Fisher
Information matrix corresponding to the treatment estimate. As shown in the
Details of \code{\link{.get_b22}}, the estimating equations for a generalized
linear model can be written as \deqn{\psi_{i} = w_{i}(y_{i} - \mu(\beta'x_{i}))x_{i}
/ \phi} The information matrix A22 is then the negative
Jacobian of \eqn{\psi_{i}}: \deqn{A_{22} = \partial\psi_{i}/\partial\beta =
-w_{i}(d\mu(x_{i}\beta)/dx_{i}\beta)x_{i}x_{i}' / \phi} In
matrix form, this can be expressed as \eqn{-X'WX}, where W is a diagonal
matrix of the vector of elementwise products of the terms not
found in the model matrix.
}
\keyword{internal}
