% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SandwichLayerVariance.R
\name{.get_b22}
\alias{.get_b22}
\title{(Internal) Get the B22 block of the sandwich variance estimator}
\usage{
.get_b22(x, ...)
}
\arguments{
\item{x}{A DirectAdjusted model object}

\item{...}{Arguments to be passed to sandwich::meatCL}
}
\value{
A (p+1)x(p+1) matrix where the dimensions are given by the number of
terms in the treatment model (p) and an Intercept term. Typically, this should
be a 2x2 matrix given the dichotomous handling of treatment variables in this
package and the use of the covariance model to offer the covariance adjustment.
}
\description{
(Internal) Get the B22 block of the sandwich variance estimator
}
\details{
This block refers to a clustered variance estimate of the treatment
effect estimate. The \code{stats} package offers family objects with
canonical link functions, so the log-likelihood for a generalized linear model
can be written in terms of the linear predictor as \deqn{L(yi, \beta, \phi, wi)
= wi * (yi * \beta'xi - b(\beta'xi)) / \phi + h(yi; \phi)} The estimating equations
\eqn{\psii} given by the score function can then be expressed as
\deqn{\psii = E[wi * (yi - \mu(\beta'xi)) * xi / \phi]}
In section 4.4 of the second edition of Categorical Data Analysis, Agresti
shows the derivative of the mean function with respect to the linear predictor
is equivalent to the weighted variance for an observation divided by the estimate
of the dispersion parameter. Thus, the above estimating equations can also be
written as \deqn{\psii = (ri / Var(yi)) * (d\mui/d\etai) * xi}

A matrix C of dimension nxJ is formed to indicate which unit each subject in
the design belongs to, where J is the number of units at the level of treatment
assignment. The treatment assignment-level estimating equations are then obtained
via \eqn{C'\psi}, where \eqn{\psi} is the matrix of estimating equations at the
subject level.
}
\references{
Agresti, Alan. Categorical Data Analysis. 2003. Open WorldCat,
https://nbn-resolving.org/urn:nbn:de:101:1-201502241089.
}
\keyword{internal}
