% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SandwichLayer.R
\name{.get_ca_and_prediction_gradient}
\alias{.get_ca_and_prediction_gradient}
\title{(Internal) Get the a vector of "response" predictions from a covariance model
and its gradient with respect to the fitted coefficients}
\usage{
.get_ca_and_prediction_gradient(model, newdata = NULL)
}
\arguments{
\item{model}{Any model that inherits from a \code{glm}, \code{lm}, or \code{
robustbase::lmrob} object}

\item{newdata}{Optional; a data.frame of new data}
}
\value{
Covariate adjusted outcomes and their prediction gradient (a list of
a numeric vector and a matrix)
}
\description{
(Internal) Get the a vector of "response" predictions from a covariance model
and its gradient with respect to the fitted coefficients
}
