% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SandwichLayerVariance.R
\name{.get_dof}
\alias{.get_dof}
\title{Get the degrees of freedom of a sandwich variance estimate associated with a teeMod fit}
\usage{
.get_dof(x, vcov_type, ell, cluster = NULL, cls = NULL, ...)
}
\description{
Get the degrees of freedom of a sandwich variance estimate associated with a teeMod fit
}
\keyword{internal}
