% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DirectAdjusted.R
\name{.make_uoa_ids}
\alias{.make_uoa_ids}
\title{Make unit of assignment ID's to pass to \code{sandwich::meatCL} \code{cluster}
argument}
\usage{
.make_uoa_ids(x, cluster = NULL, ...)
}
\arguments{
\item{x}{a fitted \code{DirectAdjusted} object}

\item{cluster}{character vector or list; optional. Specifies column names that appear
in both the covariance adjustment dataframe C and quasiexperimental dataframe
Q. Defaults to NULL, in which case unit of assignment columns indicated in
the Design will be used to generate clustered covariance estimates.}

\item{...}{arguments passed to methods}
}
\value{
A vector of length \eqn{|Q| + |C} \ \eqn{Q|}
}
\description{
Make unit of assignment ID's to pass to \code{sandwich::meatCL} \code{cluster}
argument
}
\details{
These ID's should align with the output of \code{estfun.DirectAdjusted},
which stacks the rows from Q atop the rows in C that don't overlap with Q.
}
\keyword{internal}
