% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SandwichLayerVariance.R
\name{.make_uoa_ids}
\alias{.make_uoa_ids}
\title{Make unit of assignment ID's that align with the output of \code{estfun.DirectAdjusted}}
\usage{
.make_uoa_ids(x, cluster = NULL, ...)
}
\description{
Make unit of assignment ID's that align with the output of \code{estfun.DirectAdjusted}
}
\keyword{internal}
