% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SandwichLayer.R
\name{.sanitize_C_ids}
\alias{.sanitize_C_ids}
\title{(Internal) Return ID's used to order observations in the covariance
adjustment sample}
\usage{
.sanitize_C_ids(x, id_col = NULL, sorted = FALSE, ...)
}
\arguments{
\item{x}{a \code{SandwichLayer} object}

\item{id_col}{character vector or list; optional. Specifies column names that
appear in botn the covariance adjustment and direct adjustmet samples.
Defaults to NULL, in which case unit of assignment columns in the
\code{SandwichLayer}'s \code{StudySpecification} slot will be used to
generate ID's.}

\item{...}{arguments passed to methods}
}
\value{
A vector of length equal to the number of units of observation used
to fit the covariance adjustment model
}
\description{
(Internal) Return ID's used to order observations in the covariance
adjustment sample
}
\keyword{internal}
