% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DirectAdjusted.R
\name{.sanitize_Q_ids}
\alias{.sanitize_Q_ids}
\title{Generate a list of sanitized units of assignment from Q}
\usage{
.sanitize_Q_ids(x, by = NULL, sorted = FALSE, ...)
}
\arguments{
\item{x}{a fitted \code{DirectAdjusted} object}

\item{by}{Defaults to NULL, which means unit of assignment columns
indicated in the Design will be used to align contributions to estimating
equations or generate clustered covariance estimates. A non-NULL argument to
\code{by} specifies a string or character vector of column names appearing in both
the covariance adjustment and quasiexperimental sa,ples that should be used
for clustering covariance estimates.}

\item{sorted}{Defaults to FALSE, which provides the unit of assignment ID's
in the same order as the model frame. If TRUE, the unit of assignment ID's
will be sorted alphanumerically and returned as a list, as typical of the
output of \code{sort.int} when \code{index.return = TRUE}.}

\item{...}{arguments passed to methods}
}
\value{
If not \code{sorted}, a vector of length \eqn{|Q|}, where Q is the
quasiexperimental sample. If \code{sorted}, a list whose elements are vectors of
length \eqn{\{|i : i \in Q|\}}. The sorted output will be used to align observations'
contributions to the ITT effect model with their contributions to the covariance
adjustment model in \code{estfun.DirectAdjusted}.
}
\description{
Generate a list of sanitized units of assignment from Q
}
\keyword{internal}
