% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DirectAdjusted.R
\name{.sanitize_Q_ids}
\alias{.sanitize_Q_ids}
\title{Return ID's for observations in \eqn{Q}}
\usage{
.sanitize_Q_ids(x, id_col = NULL, ...)
}
\arguments{
\item{x}{a fitted \code{DirectAdjusted} object}

\item{id_col}{character vector; optional. Specifies column(s) whose ID's will
be returned. The column must exist in the data that created the \code{Design}
object. Default is NULL, in which case unit of assignment columns indicated
in the design will be used to generate ID's.}

\item{...}{arguments passed to methods}
}
\value{
A vector of length \eqn{|Q|}
}
\description{
Return ID's for observations in \eqn{Q}
}
\keyword{internal}
