% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teeMod.R
\name{.sanitize_Q_ids}
\alias{.sanitize_Q_ids}
\title{(Internal) Return ID's used to order observations in the direct
adjustment sample}
\usage{
.sanitize_Q_ids(x, id_col = NULL, ...)
}
\arguments{
\item{x}{a fitted \code{teeMod} model}

\item{id_col}{character vector; optional. Specifies column(s) whose ID's will
be returned. The column must exist in the data that created the
\code{StudySpecification} object. Default is NULL, in which case unit of
assignment columns indicated in the specification will be used to generate
ID's.}

\item{...}{arguments passed to methods}
}
\value{
A vector with length equal to the number of units of observation in
the direct adjustment sample
}
\description{
(Internal) Return ID's used to order observations in the direct
adjustment sample
}
\keyword{internal}
