% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StudySpecificationAccessors.R
\name{.update_call_formula}
\alias{.update_call_formula}
\title{(Internal) Updates \code{spec@call}'s formula with the currently defined
variable names.}
\usage{
.update_call_formula(specification)
}
\arguments{
\item{specification}{A \code{StudySpecification}}
}
\value{
An updated formula
}
\description{
Helper function to update the \code{call} with the appropriate variable
names after they've been modified. Called within \code{StudySpecification}
replacers.
}
\details{
It's return should be stuck into the specification via
\code{spec@call$formula <- .update_call_formula(spec)}
}
\keyword{internal}
