% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StudySpecificationAccessors.R
\name{.update_structure}
\alias{.update_structure}
\title{(Internal) Replaces \code{type} columns in \code{specification} with
\code{new}}
\usage{
.update_structure(specification, new, type)
}
\arguments{
\item{specification}{A \code{StudySpecification}}

\item{new}{A named \code{data.frame} with the replacement, should be the
output of \code{.convert_to_data.frame()}.}

\item{type}{One of "t", "f", "u" or "b\".}
}
\value{
The updated \code{StudySpecification}
}
\description{
Assumes \code{.convert_to_data.frame()} has already been called on
\code{new}
}
\keyword{internal}
