% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DirectAdjusted.R
\name{estfun.DirectAdjusted}
\alias{estfun.DirectAdjusted}
\title{Extract empirical estimating equations from a \code{DirectAdjusted} model fit}
\usage{
\method{estfun}{DirectAdjusted}(x, ...)
}
\arguments{
\item{x}{a fitted \code{DirectAdjusted} object}

\item{...}{arguments passed to methods}
}
\value{
An \eqn{n\times k} matrix of empirical estimating equations for the
covariance-adjusted ITT effect regression. \eqn{n} represents the number of
observations in the union of the samples used to fit the two regressions.
\eqn{k} represents the number of parameters in the latter model.\cr\cr
Each row represents an observation's contribution to the stacked estimating
equations. This contribution, denoted \eqn{\tilde{\psi}_{i}} for the \eqn{i}th
observation, is given by \deqn{\tilde{\psi}_{i} = \psi_{i} +
\phi_{i}A_{11}^{-1}A_{21}^{T}} where \eqn{\psi_{i}} is the observation's
contribution to the ITT effect model fit, \eqn{\phi_{i}} is the observation's
contribution to the covariance adjustment model fit, and the \eqn{A} matrices
are given by typical sandwich calculations. The output matrix is orderded
such that the units used to fit the latter comprise the initial rows, and
any additional observations used to fit the former are stacked below.\cr\cr
Note that the formulation of the output matrix \eqn{\tilde{\Psi}} requires
information about each observation's contributions to both the covariance
adjustment and ITT effect models (where some observations may not contribute
to both models). Estimating equations are taken from \code{sandwich::estfun}
calls on both fitted models and aligned as closely as possible.
The \code{by} argument in \code{cov_adj()} can be used to specify a column unrelated to
the design that will allow for exact alignment of such matrices. If no \code{by}
argument is provided, clustering information given in the \code{DirectAdjusted}'s
\code{Design} object will be used to align rows by unit of assignment, even
though no guarantees can be made about aligning the matrices within units
of assignment. Regardless of the eventual alignment and initial ordering of the
observations in the two matrices, however, when using \code{vcovDA},
variance estimates will ultimately be the same due to the clustering passed
to any \code{sandwich::meatCL} calls.
}
\description{
Extract empirical estimating equations from a \code{DirectAdjusted} model fit
}
