% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SandwichLayerVariance.R
\name{get_da_bread}
\alias{get_da_bread}
\title{Get the "bread" matrix for the Direct Adjustment Model}
\usage{
get_da_bread(x)
}
\arguments{
\item{x}{DirectAdjusted}
}
\value{
scalar
}
\description{
The "bread" matrix of the direct adjustment model is the inverse
of the model's Fisher information matrix when the gradient is taken with
respect to the treatment variable
}
