% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SandwichLayerVariance.R
\name{get_overlap_vcov_matrix}
\alias{get_overlap_vcov_matrix}
\title{Generate Cluster-level Estimating Equations Covariance Matrix
for Overlapping Rows in Experimental and Covariance Model Data}
\usage{
get_overlap_vcov_matrix(x)
}
\arguments{
\item{x}{DirectAdjusted}
}
\value{
outer product of cluster-level estimating equations for the covariance model
and the treatment, matrix should have dimension p x k where p is the column
count in the covariance model matrix and k is the column count in the treatment
model
}
\description{
\code{get_overlap_vcov_matrix} returns the covariance matrix of the
cluster-level estimating equations for rows that appear in both the experimental
and covariance model data
}
