% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StudySpecificationUtilities.R
\name{has_binary_treatment}
\alias{has_binary_treatment}
\title{Check whether treatment stored in a \code{StudySpecification} object
is binary}
\usage{
has_binary_treatment(spec)
}
\arguments{
\item{spec}{\code{StudySpecification} object}
}
\value{
logical vector of length 1
}
\description{
Check whether treatment stored in a \code{StudySpecification} object
is binary
}
