% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DesignUtilities.R
\name{identical_Designs}
\alias{identical_Designs}
\title{Check if \code{Design} objects are identical}
\usage{
identical_Designs(x, y, dichotomy_force = FALSE)
}
\arguments{
\item{x}{A \code{Design} object.}

\item{y}{A \code{Design} object.}

\item{dichotomy_force}{Logical, default \code{FALSE}. If \code{FALSE}, the
\code{dichotomy()} of \code{x} and \code{y} is ignored in the comparison.
(In other words, two \code{Design}s which differ only in their
\code{@dichotomy} slot will be considered identical.) If \code{TRUE}, the
\code{dichotomy} must also be in agreement between \code{x} and \code{y}.}
}
\value{
Logical, are \code{x} and \code{y} identical?
}
\description{
Check if \code{Design} objects are identical
}
