% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DesignUtilities.R
\name{identify_small_blocks}
\alias{identify_small_blocks}
\title{Identify fine strata (blocks with one treated or one control unit of assignment)}
\usage{
identify_small_blocks(des)
}
\arguments{
\item{des}{A \code{Design} object.}
}
\value{
Logical vector with length given by the number of blocks in the
Design
}
\description{
Identify fine strata (blocks with one treated or one control unit of assignment)
}
