% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StudySpecificationUtilities.R
\name{identify_small_blocks}
\alias{identify_small_blocks}
\title{Identify fine strata}
\usage{
identify_small_blocks(spec)
}
\arguments{
\item{spec}{A \code{StudySpecification} object.}
}
\value{
Logical vector with length given by the number of blocks in
\code{StudySpecification}
}
\description{
Identify blocks in a \code{StudySpecification} with exactly one
treated or one control unit of assignment.
}
