% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ittestimate.R
\name{ittestimate}
\alias{ittestimate}
\title{Estimate Treatment Effect}
\usage{
ittestimate(
  design,
  data,
  outcome,
  target = ate,
  covAdjModel = NULL,
  clusterIds = NULL,
  ...
)
}
\arguments{
\item{design}{study Design}

\item{data}{Data for analysis}

\item{outcome}{string containing name of outcome variable in \code{data}}

\item{target}{a function returning a WeightedDesign object, or either "ate"
(default) or "ett"}

\item{covAdjModel}{optional; covariable adjustment model}

\item{clusterIds}{optional; list connecting names of cluster variables in
\code{design} to cluster variables in \code{data}}

\item{...}{Additional arguments to the model.}
}
\value{
Estimated treatment effect
}
\description{
Estimate Treatment Effect
}
