% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmitt.formula.R
\docType{class}
\name{lmitt.formula-class}
\alias{lmitt.formula-class}
\title{lmitt.formula class
Class for lmitt formulas}
\description{
lmitt.formula class
Class for lmitt formulas
}
\details{
This has a specific \code{model.frame()} method so that \code{lmitt()} fits
the model we desire
}
