% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmitt.R
\name{lmitt}
\alias{lmitt}
\title{Linear Model for Intention To Treat}
\usage{
lmitt(formula, design = NULL, target = NULL, ...)
}
\arguments{
\item{formula}{See \code{lm()}}

\item{design}{Optional, explicitly specify the \code{Design} to be used. If
the \code{Design} is specified elsewhere in the model (e.g. passed as an
argument to any of \code{ate()}, \code{ett()}, \code{cov_adj()} or
\code{adopters()}) it will be found automatically and does not need to be
passed here as well. (If the \code{Design} is found in the model, this
argument is ignored.)}

\item{target}{Optional, explicitly specify the estimand. If the
\code{weights} are generated using either \code{ate()} or \code{ett()},
the \code{target]} will be found automatically. Otherwise, specify whether
the goal is estimating ATE ("ate") or ETT ("ett"). (If weights are
specified, this argument is ignored.)}

\item{...}{Additional arguments passed to \code{lm()}.}
}
\value{
\code{Lmitted} model.
}
\description{
Linear Model for Intention To Treat
}
