% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmitt.R
\name{lmitt}
\alias{lmitt}
\alias{lmitt.formula}
\alias{lmitt.lm}
\title{Linear Model for Intention To Treat}
\usage{
lmitt(obj, design, data, ...)

\method{lmitt}{formula}(obj, design, data, absorb = FALSE, offset = NULL, weights = NULL, ...)

\method{lmitt}{lm}(obj, design = NULL, ...)
}
\arguments{
\item{obj}{A \code{formula} or a \code{lm} object. See details.}

\item{design}{The \code{Design} to be used. Alternatively, a formula
creating a design (of the type of that would be passed as the first
argument to \code{rd_design()}, \code{rct_design()}, or
\code{obs_design()}). If the formula includes a \code{forcing()} element,
an RD design is created. Otherwise an observational design is created. An
RCT design must be created using \code{rct_design()}.}

\item{data}{Data frame such as would be passed into \code{lm()}.}

\item{...}{Additional arguments passed to \code{lm()}.}

\item{absorb}{If \code{TRUE}, fixed effects are included for units of
assignemnt/clusters identified in the \code{Design}. Excluded in
\code{FALSE}. Default is \code{FALSE}.}

\item{offset}{Offset of the kind which would be passed into \code{lm()}. To
utilize flexida's functionality, the output of \code{cov_adj()} should be
used.}

\item{weights}{Weight of the kind which would be passed into \code{lm()}. To
utilize flexida's functionality, the output of \code{ate()} or
\code{ett()}, or the strings \code{"ate"}/\code{"ett"}, should be used..}
}
\value{
\code{DirectAdjusted} model.
}
\description{
Generates a linear model object which allows proper estimation of variances
accounting for the study design.
}
\details{
The first argument to \code{lmitt()} should be a formula specifying the
outcome on the left hand side. The right hand side should be either one of
"1", requesting a singluar treatment effect, or a subgrouping variable,
requesting a treatment effect be estimated for each subgroup.

Alternatively, \code{obj} can be a pre-created \code{lm} object. No
modification is made to the formula of the object. See the help for
\code{as.lmitt()} for details of this conversion.

Note that although the \code{Design} creation functions (e.g.
\code{rct_design()}) take an optional \code{subset=} argument used in the
creation of the \code{Design}, this is \bold{not} the same as the
\code{subset=} argument passed to \code{lm()} or \code{lmitt()}. The
\code{subset=} argument when creating a \code{Design} restricts the data
used to generate the \code{Design}, but has no direct impact on the future
\code{lm()} or \code{lmitt()} calls using that \code{Design}. (It can
indirectly have an impact by excluding particular clusters/units of
assignment from recieving a treatment assignment and thus complete case
analysis removes them from the model.)

On the other hand, the \code{subset=} argument in \code{lm()} or
\code{lmitt()} refers only to subsetting the \code{data} argument passed
into \code{lm()} or \code{lmitt()}.

\code{lmitt()} will produce a message if the \code{design} passed in has
block information that is not being utilized in the model. Note that this is
\emph{not} an error, but could be an oversight. To disable this message, run
\code{options("flexida_message_on_unused_blocks" = FALSE)}.
}
