% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmitt.R
\name{lmitt}
\alias{lmitt}
\title{Linear Model for Intention To Treat}
\usage{
lmitt(formula, design = NULL, ...)
}
\arguments{
\item{formula}{See \code{lm()}}

\item{design}{Optional, explicitly specify the \code{Design} to be used. If
the \code{Design} is specified elsewhere in the model (e.g. passed as an
argument to any of \code{ate()}, \code{ett()}, \code{cov_adj()} or
\code{adopters()}) it will be found automatically and does not need to be
passed here as well. (If different \code{Design} objects are passed
(either through the \code{lm} in weights or covariance adjustment, or
through this argument), an error will be produced.)}

\item{...}{Additional arguments passed to \code{lm()}.}
}
\value{
\code{Lmitted} model.
}
\description{
Linear Model for Intention To Treat
}
