% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{michigan_school_pairs}
\alias{michigan_school_pairs}
\title{Intervention data from a pair-matched study of schools in Michigan}
\format{
A \code{data.frame} with 14 rows and 3 columns.
}
\usage{
michigan_school_pairs
}
\description{
Schools matched into pairs or triplets approximating the subset
of participating schools in Pane et al. (2014) from Michigan
}
\references{
Pane, John F., et al. "Effectiveness of cognitive tutor algebra I at scale."
\emph{Educational Evaluation and Policy Analysis} 36.2 (2014): 127-144.
}
\keyword{dataset}
