% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SandwichLayer.R
\name{show,PreSandwichLayer-method}
\alias{show,PreSandwichLayer-method}
\title{Show a \code{PreSandwichLayer} or \code{SandwichLayer}}
\usage{
\S4method{show}{PreSandwichLayer}(object)
}
\arguments{
\item{object}{\code{PreSandwichLayer} or \code{SandwichLayer} object}
}
\value{
an invisible copy of \code{object}
}
\description{
Display information about a \code{PreSandwichLayer} or
\code{SandwichLayer} object
}
