% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SandwichLayer.R
\name{subset,PreSandwichLayer-method}
\alias{subset,PreSandwichLayer-method}
\alias{[,PreSandwichLayer-method}
\title{\code{PreSandwichLayer} and \code{SandwichLayer} subsetting}
\usage{
\S4method{subset}{PreSandwichLayer}(x, subset)

\S4method{[}{PreSandwichLayer}(x, i)
}
\arguments{
\item{x}{\code{PreSandwichLayer} or \code{SandwichLayer} object}

\item{subset}{Logical vector identifying values to keep or drop}

\item{i}{indices specifying elements to extract or replace. See
\code{help("[")} for further details.}
}
\value{
\code{x} subset by \code{subset} or \code{i}
}
\description{
Return subset of a \code{PreSandwichLayer} or
\code{SandwichLayer} which meets conditions.
}
