% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.Design.R
\name{print.summary.Design}
\alias{print.summary.Design}
\title{Print summary object}
\usage{
\method{print}{summary.Design}(x, ..., max_unit_print = 3)
}
\arguments{
\item{x}{Design object}

\item{...}{Other args}

\item{max_unit_print}{Maximum number of treatment levels to print in treatment
table}
}
\value{
object, invisibly
}
\description{
Print summary object
}
