% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.DirectAdjusted.R
\name{print.summary.DirectAdjusted}
\alias{print.summary.DirectAdjusted}
\title{Print summary of \code{DirectAdjusted} object}
\usage{
\method{print}{summary.DirectAdjusted}(x, ..., max_unit_print = 3)
}
\arguments{
\item{x}{\code{summary.DirectAdjusted} object}

\item{...}{Other args}

\item{max_unit_print}{Maximum number of treatment levels to print in
treatment table}
}
\value{
object, invisibly
}
\description{
Print summary of \code{DirectAdjusted} object
}
