% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design.R
\name{show,Design-method}
\alias{show,Design-method}
\title{Show a \code{Design}}
\usage{
\S4method{show}{Design}(object)
}
\arguments{
\item{object}{\code{Design} object}
}
\value{
an invisible copy of \code{object}
}
\description{
Show a \code{Design}
}
