% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design.R
\name{show,Design-method}
\alias{show,Design-method}
\title{Display a \code{Design}}
\usage{
\S4method{show}{Design}(object)
}
\arguments{
\item{object}{a \code{Design} object}
}
\value{
an invisible copy of \code{object}
}
\description{
Display a \code{Design}
}
