% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DirectAdjusted.R
\name{show,DirectAdjusted-method}
\alias{show,DirectAdjusted-method}
\title{Show an DirectAdjusted}
\usage{
\S4method{show}{DirectAdjusted}(object)
}
\arguments{
\item{object}{DirectAdjusted object}
}
\value{
an invisible copy of \code{object}
}
\description{
Show an DirectAdjusted
}
