% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StudySpecification.R
\name{show,StudySpecification-method}
\alias{show,StudySpecification-method}
\title{Show a \code{StudySpecification}}
\usage{
\S4method{show}{StudySpecification}(object)
}
\arguments{
\item{object}{\code{StudySpecification} object, usually a result of a call
to \code{\link[=rct_spec]{rct_spec()}}, \code{\link[=obs_spec]{obs_spec()}}, or \code{\link[=rd_spec]{rd_spec()}}.}
}
\value{
\code{object}, invisibly.
}
\description{
Display information about a \code{StudySpecification} object
}
