% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teeMod.R
\name{show,teeMod-method}
\alias{show,teeMod-method}
\title{Show a \code{teeMod}}
\usage{
\S4method{show}{teeMod}(object)
}
\arguments{
\item{object}{\code{teeMod} object, usually a result of a call to
\code{\link[=lmitt]{lmitt()}}.}
}
\value{
\code{object}, invisibly.
}
\description{
Display information about a \code{teeMod} object
}
\examples{
library(sandwich)
library(propertee)
data(schooldata)
data(studentdata)
spec <- rct_spec(treatment ~ unitid(schoolid), schooldata)
show(spec)
summary(spec)
tm <- lmitt(math ~ 1, spec, studentdata, weights = "ate")
show(tm)
summary(tm)
bread(tm)
ef <- estfun(tm)
head(ef)
covadj_mod <- lm(math ~ gpa, studentdata)
tm <- lmitt(math ~ 1, spec, studentdata, weights = "ate",
            offset = cov_adj(covadj_mod))
bread(tm)
ef <- estfun(tm)
head(ef)
vcov(tm)
vcov(tm, type = "HC1")
vcov(tm, cov_adj_rcorrect = "HC1")
confint(tm, 2)
}
