% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeightedDesign.R
\name{show,WeightedDesign-method}
\alias{show,WeightedDesign-method}
\title{Show a WeightedDesign}
\usage{
\S4method{show}{WeightedDesign}(object)
}
\arguments{
\item{object}{WeightedDesignDesign object}
}
\value{
an invisible copy of \code{object}
}
\description{
Show a WeightedDesign
}
