% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeightedStudySpecification.R
\name{show,WeightedStudySpecification-method}
\alias{show,WeightedStudySpecification-method}
\title{Show a \code{WeightedStudySpecification}}
\usage{
\S4method{show}{WeightedStudySpecification}(object)
}
\arguments{
\item{object}{a \code{WeightedStudySpecification} object}
}
\value{
an invisible copy of \code{object}
}
\description{
Prints out the weights from a \code{WeightedStudySpecification}
}
\examples{
data(schooldata)
data(studentdata)
spec <- rct_spec(treatment ~ unitid(schoolid) + block(state), schooldata)
wt <- ate(spec, data = schooldata)
wt.subset <- wt[seq_len(4)]
identical(subset(wt, c(rep(TRUE, 4), rep(FALSE, nrow(schooldata)-4))),
          wt.subset)
show(wt.subset)
}
