% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeightedStudySpecification.R
\name{show,WeightedStudySpecification-method}
\alias{show,WeightedStudySpecification-method}
\title{Show a \code{WeightedStudySpecification}}
\usage{
\S4method{show}{WeightedStudySpecification}(object)
}
\arguments{
\item{object}{a \code{WeightedStudySpecification} object}
}
\value{
an invisible copy of \code{object}
}
\description{
Prints out the weights from a \code{WeightedStudySpecification}
}
