% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simdata}
\alias{simdata}
\title{Simulated data}
\format{
A data.frame with 100 rows and 7 columns.
\itemize{
\item \code{uoa1} First level unit of assignment ID
\item \code{uoa2} Second level unit of assignment ID
\item \code{bid} Block ID
\item \code{force} Forcing variable
\item \code{z} Binary treatment indicator
\item \code{o} 4-level ordered treatment variable
\item \code{dose} Dose treatment variable
\item \code{x} Some predictor
\item \code{y} Some outcome
}
}
\usage{
simdata
}
\description{
Simulated data to use with the \bold{propertee} package with unit
of assignment level treatment assignment
}
\keyword{dataset}
