% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{STARdata}
\alias{STARdata}
\title{STAR data}
\format{
A \code{data.frame} with 11,598 rows and 48 columns.
}
\usage{
STARdata
}
\description{
Data from Tennessee’s Project STAR study.
This is a copy of \code{AER::STAR}, describing student
participants in the Project STAR (Student-Teacher Achievement
Ratio) study conducted in Tennessee, USA in the late 1980s.
Variables are as described in the \code{AER} package's \code{STAR}
help page, with the exception that the \code{row.names}
of that data frame have been moved into a new first column,
\code{studentid}.
}
\keyword{dataset}
