% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.StudySpecification.R
\name{summary.StudySpecification}
\alias{summary.StudySpecification}
\alias{print.summary.StudySpecification}
\title{Summarizing \code{StudySpecification} objects}
\usage{
\method{summary}{StudySpecification}(object, ..., treatment_binary = TRUE)

\method{print}{summary.StudySpecification}(x, ..., max_unit_print = 3)
}
\arguments{
\item{object}{\code{StudySpecification} object, usually a result of a call
to \code{\link[=rct_spec]{rct_spec()}}, \code{\link[=obs_spec]{obs_spec()}}, or \code{\link[=rd_spec]{rd_spec()}}.}

\item{...}{Ignored}

\item{treatment_binary}{Should the treatment be dichotomized if
\code{object} contains a \code{dichotomy}? Ignored if \code{object} does
not contain a \code{dichotomy}.}

\item{x}{\code{summary.StudySpecification} object, usually as a result of a
call to \code{\link[=summary.StudySpecification]{summary.StudySpecification()}}}

\item{max_unit_print}{Maximum number of treatment levels to print in
treatment table}
}
\value{
The \code{StudySpecification} or
\code{summary.StudySpecification}object, invisibly
}
\description{
\code{\link[=summary]{summary()}} method for class \code{StudySpecification}.
}
