% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.Design.R
\name{summary.Design}
\alias{summary.Design}
\title{Create summary of \code{Design} object}
\usage{
\method{summary}{Design}(object, ...)
}
\arguments{
\item{object}{Design object}

\item{...}{Other args}
}
\value{
object of class \code{summary.Design}
}
\description{
Create summary of \code{Design} object
}
