% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.Design.R
\name{summary.Design}
\alias{summary.Design}
\title{Summary of Design object}
\usage{
\method{summary}{Design}(object, ...)
}
\arguments{
\item{object}{Design object}

\item{...}{Other args}
}
\value{
object, invisibly
}
\description{
Summary of Design object
}
