% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.DirectAdjusted.R
\name{summary.DirectAdjusted}
\alias{summary.DirectAdjusted}
\title{Create summary of \code{DirectAdjusted} object}
\usage{
\method{summary}{DirectAdjusted}(object, vcov.type = "CR0", ...)
}
\arguments{
\item{object}{DirectAdjusted object}

\item{vcov.type}{A string indicating the desired variance estimator. See
help for \code{vcovDA} for details on accepted types.}

\item{...}{Additional arguments to \code{vcovDA()}, such as the desired
finite sample heteroskedasticity-robust standard error adjustment.}
}
\value{
object of class \code{summary.DirectAdjusted}
}
\description{
If a \code{DirectAdjusted} object is fit with a \code{SandwichLayer} offset,
then the usual \code{stats::summary.lm()} output is enhanced by the use of
covariance-adjusted sandwich standard errors, with t-test values
recalculated to reflect the new standard errors.
}
