% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design.R
\name{treatment_table}
\alias{treatment_table}
\title{Generate table of number of units/clusters within each treatment
level}
\usage{
treatment_table(design, ...)
}
\arguments{
\item{design}{A Design object}

\item{...}{add'l optional arguments to \code{table}}
}
\value{
a table of treatment by units
}
\description{
Returns a table of number of units of assignment in each treatment group,
sorted by the size of the groups
}
\examples{
des <- rd_design(z ~ uoa(cid1, cid2) + forcing(force), data = simdata)
treatment_table(des)
}
