% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SandwichLayerVariance.R
\name{vcovDA}
\alias{vcovDA}
\alias{.vcov_CR0}
\alias{.vcov_HC0}
\alias{.vcov_MB0}
\title{Compute covariance-adjusted cluster-robust sandwich variance estimates}
\usage{
vcovDA(x, type = "CR0", cluster = NULL, ...)

.vcov_CR0(x, ...)

.vcov_HC0(x, ...)

.vcov_MB0(x, ...)
}
\arguments{
\item{x}{a fitted \code{DirectAdjusted} model object}

\item{type}{A string indicating the desired variance estimator. Currently
accepts "CR0", "MB0", or "HC0"}

\item{cluster}{Defaults to NULL, which means unit of assignment columns
indicated in the Design will be used to generate clustered covariance estimates.
A non-NULL argument to \code{cluster} specifies a string or character vector of
column names appearing in both the covariance adjustment and quasiexperimental
samples that should be used for clustering covariance estimates.}

\item{...}{Arguments to be passed to the internal variance estimation function.}
}
\value{
A \eqn{2\times 2} matrix where the dimensions are
given by the intercept and treatment variable terms in the ITT effect model
}
\description{
Compute covariance-adjusted cluster-robust sandwich variance estimates

Model-based standard errors with HC0 adjustment
}
\details{
Supported \code{type} include:
\itemize{
\item \code{"CR0"}, \code{"MB0"}, \code{"HC0"} are synonyms for ...
\item Others...
}
}
\keyword{internal}
