% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SandwichLayerVariance.R
\name{vcovDA}
\alias{vcovDA}
\alias{.vcovMB_CR0}
\title{Compute covariance-adjusted cluster-robust sandwich variance estimates}
\usage{
vcovDA(x, type = c("CR0"), ...)

.vcovMB_CR0(x, ...)
}
\arguments{
\item{x}{a fitted \code{DirectAdjusted} model object}

\item{type}{A string indicating the desired variance estimator. Currently
accepts "CR0"}

\item{...}{Arguments to be passed to the internal variance estimation function.
One argument a user may want to manually override is the \code{cluster} argument.
Users may be interested in clustering standard errors at levels different than
the unit of assignment level specified in the \code{DirectAdjusted} model's
\code{Design}. In this case, they may specify column names in the ITT effect/covariance
adjustment model datasets corresponding to a different clustering level.
\code{.get_b11()} and \code{.get_b12()} tolerate NA values for manually
provided cluster ID's, since the covariance adjustment model may be fit to a
sample larger than the quasiexperimental sample. In this case, each unit with
an NA cluster ID will be treated as an independent observation. \code{.get_b22()},
however, will fail if the new cluster levels contain NA's because all units
in the quasiexperimental sample should have well-defined design information.}
}
\value{
A \eqn{2\times 2} matrix where the dimensions are
given by the intercept and treatment variable terms in the ITT effect model
}
\description{
Compute covariance-adjusted cluster-robust sandwich variance estimates

Model-based standard errors with HC0 adjustment
}
\keyword{internal}
