% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SandwichLayerVariance.R
\name{vcovDA}
\alias{vcovDA}
\alias{.vcovMB_CR0}
\title{Compute covariance-adjusted cluster-robust sandwich variance estimates}
\usage{
vcovDA(x, type = c("CR0"), cluster = NULL, ...)

.vcovMB_CR0(x, ...)
}
\arguments{
\item{x}{a fitted \code{DirectAdjusted} model object}

\item{type}{A string indicating the desired variance estimator. Currently
accepts "CR0"}

\item{cluster}{Defaults to NULL, which means unit of assignment columns
indicated in the Design will be used to generate clustered covariance estimates.
A non-NULL argument to \code{cluster} specifies a string or character vector of
column names appearing in both the covariance adjustment and quasiexperimental
samples that should be used for clustering covariance estimates.}

\item{...}{Arguments to be passed to the internal variance estimation function.}
}
\value{
A \eqn{2\times 2} matrix where the dimensions are
given by the intercept and treatment variable terms in the ITT effect model
}
\description{
Compute covariance-adjusted cluster-robust sandwich variance estimates

Model-based standard errors with HC0 adjustment
}
\keyword{internal}
