% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SandwichLayerVariance.R
\name{vcov_tee}
\alias{vcov_tee}
\alias{.vcov_DB0}
\title{Variance/Covariance for \code{teeMod} objects}
\usage{
vcov_tee(x, type = "CR0", cluster = NULL, ...)

.vcov_DB0(x, ...)
}
\arguments{
\item{x}{a fitted \code{teeMod} model}

\item{type}{a string indicating the desired variance estimator. See Details
for supported variance estimators}

\item{cluster}{a string or character vector of column names indicating
columns to cluster standard errors by. With prior covariance adjustment,
columns must appear in both the covariance adjustment and direct adjustment
samples. Default is NULL, which uses unit of assignment columns in the
\code{StudySpecification} slot of the \code{teeMod} model.}

\item{...}{arguments to be passed to the internal variance estimation
function.}
}
\value{
A \eqn{2\times 2} matrix corresponding to an intercept and the
treatment variable in the direct adjustment model
}
\description{
Compute robust sandwich variance estimates with optional
covariance adjustment
}
\details{
Supported \code{type} include:
\itemize{
\item \code{"MB0"}, \code{"HC0"}, and \code{"CR0"} for model-based HC0 standard errors
\item \code{"MB1"}, \code{"HC1"}, and \code{"CR1"} for model-based standard errors
with HC1 corrections based on the direct adjustment estimate i.e.,
\eqn{n/(n - 2)} for \code{"MB1"} and \code{"HC1"}, and for \code{"CR1"},
\eqn{g\cdot(n-1)/((g-1)\cdot(n-2))}, where \eqn{g} is the number of clusters
in the direct adjustment sample.
\item \code{"DB0"} for specification-based HC0 variance estimates
}

To create your own \code{type}, simply define a function \code{.vcov_XXX}.
\code{type = "XXX"} will now use your method. Your method should return a
matrix of appropriate dimension, with \code{attribute} \code{type = "XXX"}.

The specification-based variance estimates can be calculated for
\code{teeMod} models satisfying the following requirements:
\itemize{
\item The model uses \code{rct_spec} as \code{StudySpecification}
\item The model only estimates a main treatment effect
\item Inverse probability weighting is incorporated
}
}
\keyword{internal}
