% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SandwichLayerVariance.R
\name{vcovDA}
\alias{vcovDA}
\alias{.vcovMB_CR1}
\title{Compute directly adjusted cluster-robust sandwich variance estimates}
\usage{
vcovDA(x, type = c("CR1"), ...)

.vcovMB_CR1(x, ...)
}
\arguments{
\item{x}{A \code{DirectAdjusted} model}

\item{type}{A string indicating the desired variance estimator. Currently
accepts "CR1"}

\item{...}{Arguments to be passed to sandwich::meatCL}
}
\value{
\code{.vcovMB_CR1()}: A \eqn{2\times 2} matrix where the dimensions are
given by the intercept and treatment variable terms in the direct adjustment
model
}
\description{
Compute directly adjusted cluster-robust sandwich variance estimates

Compute directly-adjusted CR1 cluster-robust sandwich variance estimates
under model-based assumptions
}
\keyword{internal}
