% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design.R
\name{var_names}
\alias{var_names}
\title{Return names of variables defining the \code{Design}}
\usage{
var_names(x, type)
}
\arguments{
\item{x}{a \code{Design} object}

\item{type}{one of "t", "u", "b", "f"; for "treatment",
"unit_of_assignment", "block", and "forcing" respectively}
}
\value{
character vector of variable names of the given type
}
\description{
Return names of variables defining the \code{Design}
}
\examples{
des <- obs_design(o ~ unitid(cid1, cid2), data = simdata)
var_names(des, "t")
var_names(des, "u")
var_names(des, "b")
}
