% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DirectAdjusted.R
\name{vcov.DirectAdjusted}
\alias{vcov.DirectAdjusted}
\title{Variance-Covariance matrix of \code{DirectAdjusted} object}
\usage{
\method{vcov}{DirectAdjusted}(object, ...)
}
\arguments{
\item{object}{DirectAdjusted}

\item{...}{Additional arguments to \code{vcovDA()} or \code{stats:::vcov.lm()}.}
}
\value{
Variance-Covariance matrix
}
\description{
Variance-Covariance matrix of \code{DirectAdjusted} object
}
\details{
If a \code{DirectAdjusted} object is fit with a \code{SandwichLayer}
offset, then its \code{vcov()} method provides a sandwich estimate of the
covariance-adjusted variance-covariance matrix. Otherwise, it provides
the default OLS estimate of the matrix.
}
