% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teeMod.R
\name{vcov.teeMod}
\alias{vcov.teeMod}
\title{Compute variance-covariance matrix for fitted \code{teeMod} model}
\usage{
\method{vcov}{teeMod}(object, ...)
}
\arguments{
\item{object}{a fitted \code{teeMod} model}

\item{...}{additional arguments to \code{vcovDA()}.}
}
\value{
A \eqn{2\times 2} matrix corresponding to an intercept and the
treatment variable in the direct adjustment model
}
\description{
An S3method for \code{stats::vcov} that computes standard errors for
\code{teeMod} models using \code{vcovDA()}.
}
\details{
\code{vcov.teeMod()} wraps around \code{vcovDA()}, so additional
arguments passed to \code{...} will be passed to the \code{vcovDA()} call.
See documentation for \code{vcovDA()} for information about necessary
arguments.
}
