% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SandwichLayerVariance.R
\name{vcovDA}
\alias{vcovDA}
\title{Compute direct adjustment sandwich variance estimator}
\usage{
vcovDA(x, ...)
}
\arguments{
\item{x}{A \code{Lmitted} model}

\item{...}{Arguments to be passed to sandwich::meatCL}
}
\value{
a \eqn{(p+2)\times(p+2)} matrix where the dimensions are given by the
number of terms in the covariance model (p), a treatment term, and an
intercept term
}
\description{
Compute direct adjustment sandwich variance estimator
}
\details{
\code{vcovDA()} provides covariance-adjusted cluster-robust variance
estimates for treatment effects.
}
