% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeightedDesign.R
\name{ett}
\alias{ett}
\alias{ate}
\title{Generate Direct Adjusted Weights}
\usage{
ett(design, data = NULL)

ate(design, data = NULL)
}
\arguments{
\item{design}{a Design object created by one of \code{RCT_Design}, \code{RD_Design},
or \code{Obs_Design}.}

\item{data}{optionally the data for the analysis to be performed on. May be
excluded if these functions are included as the \code{weights} argument of a
model.}
}
\value{
a WeightedDesign object
}
\description{
Generate Direct Adjusted Weights
}
