% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{ett}
\alias{ett}
\alias{ate}
\title{Generate Direct Adjusted Weights}
\usage{
ett(design, data = NULL, varLinks = NULL)

ate(design, data = NULL, varLinks = NULL)
}
\arguments{
\item{design}{a Design object created by one of \code{RCT_Design}, \code{RD_Design},
or \code{Obs_Design}.}

\item{data}{optionally the data for the analysis to be performed on. May be
excluded if these functions are included as the \code{weights} argument of a
model.}

\item{varLinks}{optional; list connecting names of cluster/unit of
assignment variables in \code{design} to cluster/unit of assignment variables
in \code{data}. Only needed if variable names differ.}
}
\value{
a WeightedDesign object
}
\description{
Generate Direct Adjusted Weights
}
