% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{ett}
\alias{ett}
\alias{ate}
\title{Generate Direct Adjusted Weights}
\usage{
ett(design = NULL, dichotomy = NULL, by = NULL, data = NULL)

ate(design = NULL, dichotomy = NULL, by = NULL, data = NULL)
}
\arguments{
\item{design}{a \code{Design} object created by one of \code{rct_design()},
\code{rd_design()}, or \code{obs_design()}.}

\item{dichotomy}{optional; a formula defining the dichotomy of the treatment
variable if it isn't already \code{0}/\code{1}. See details of help for
\code{rct_design()} e.g. for details.}

\item{by}{optional; named vector or list connecting names of unit of
assignment/ variables in \code{design} to unit of assignment/cluster
variables in \code{data}. Names represent variables in the Design; values
represent variables in the data. Only needed if variable names differ.}

\item{data}{optionally the data for the analysis to be performed on. May be
excluded if these functions are included as the \code{weights} argument of
a model.}
}
\value{
a \code{WeightedDesign} object
}
\description{
Generate Direct Adjusted Weights
}
\details{
In a block design, Weights are generated as a function of the ratio
of the number of treated units in a block versus the total number of units
in a block.

In any blocks where that ratio is 0 or 1 (that is, all units in the block
have the same treatment status), the weights will be 0. In effect this
removes from the target population any block in which there is no basis for
estimating either means under treatment or means under control.

If block is missing for a given observation, a weight of 0 is applied.
}
