% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeightedDesign.R
\name{subset,WeightedDesign-method}
\alias{subset,WeightedDesign-method}
\alias{[,WeightedDesign,ANY,ANY,ANY-method}
\title{\code{WeightedDesign} subsetting}
\usage{
\S4method{subset}{WeightedDesign}(x, subset)

\S4method{[}{WeightedDesign,ANY,ANY,ANY}(x, i)
}
\arguments{
\item{x}{\code{WeightedDesign} object}

\item{subset}{Logical vector identifying values to keep or drop}

\item{i}{indices specifying elements to extract or replace. See
\code{help("[")} for further details.}
}
\value{
A \code{WeightedDesign} object which is a subsetted version of
\code{x}.
}
\description{
Provides functionality to subset the weights of a
\code{WeightedDesign} object.
}
