% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeightedDesign.R
\name{+,WeightedDesign,numeric-method}
\alias{+,WeightedDesign,numeric-method}
\alias{+,numeric,WeightedDesign-method}
\alias{-,WeightedDesign,numeric-method}
\alias{-,numeric,WeightedDesign-method}
\alias{*,WeightedDesign,numeric-method}
\alias{*,numeric,WeightedDesign-method}
\alias{/,WeightedDesign,numeric-method}
\alias{/,numeric,WeightedDesign-method}
\title{\code{WeightedDesign} Operations}
\usage{
\S4method{+}{WeightedDesign,numeric}(e1, e2)

\S4method{+}{numeric,WeightedDesign}(e1, e2)

\S4method{-}{WeightedDesign,numeric}(e1, e2)

\S4method{-}{numeric,WeightedDesign}(e1, e2)

\S4method{*}{WeightedDesign,numeric}(e1, e2)

\S4method{*}{numeric,WeightedDesign}(e1, e2)

\S4method{/}{WeightedDesign,numeric}(e1, e2)

\S4method{/}{numeric,WeightedDesign}(e1, e2)
}
\arguments{
\item{e1, e2}{\code{WeightedDesign} or \code{numeric} objects}
}
\value{
a \code{WeightedDesign} object
}
\description{
Algebraic operators on \code{WeightedDesign} objects and
numeric vectors. \code{WeightedDesign}s do not support addition or
subtraction.
}
\details{
These are primarily used to either combine weights via
multiplication, or to invert weights. Addition and subtraction are not
supported and will produce errors.
}
