% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeightedDesign.R
\name{+,WeightedDesign,numeric-method}
\alias{+,WeightedDesign,numeric-method}
\alias{+,numeric,WeightedDesign-method}
\alias{-,WeightedDesign,numeric-method}
\alias{-,numeric,WeightedDesign-method}
\alias{*,WeightedDesign,numeric-method}
\alias{*,numeric,WeightedDesign-method}
\alias{/,WeightedDesign,numeric-method}
\alias{/,numeric,WeightedDesign-method}
\alias{c,WeightedDesign-method}
\title{\code{WeightedDesign} Ops}
\usage{
\S4method{+}{WeightedDesign,numeric}(e1, e2)

\S4method{+}{numeric,WeightedDesign}(e1, e2)

\S4method{-}{WeightedDesign,numeric}(e1, e2)

\S4method{-}{numeric,WeightedDesign}(e1, e2)

\S4method{*}{WeightedDesign,numeric}(e1, e2)

\S4method{*}{numeric,WeightedDesign}(e1, e2)

\S4method{/}{WeightedDesign,numeric}(e1, e2)

\S4method{/}{numeric,WeightedDesign}(e1, e2)

\S4method{c}{WeightedDesign}(x, ..., force_dichotomization_equal = FALSE)
}
\arguments{
\item{e1}{\code{WeightedDesign} or numeric}

\item{e2}{numeric or \code{WeightedDesign}}

\item{x}{\code{WeightedDesign}}

\item{...}{additional \code{WeightedDesign}s}

\item{force_dichotomization_equal}{if \code{FALSE} (default), Designs are
considered equivalent even if their \code{dichotomization} differs. If
\code{TRUE}, \code{@dichotomization} must also be equal.}
}
\description{
\code{WeightedDesign}s do not support addition or subtraction, but do
support all other reasonable operations.
}
\details{
Concatenating \code{WeightedDesign}s with \code{c()} requires both
individual \code{WeightedDesign}s to come from the same \code{Design}
(except \code{dichotomization}, see below) and have the target (\code{ate()}
or \code{ett()}). Both arguments to \code{c()} must be
\code{WeightedDesign}.

One exception when concatenting \code{WeightedDesign}s with different
\code{Design} exists. There may be cases where the treatment is continuous
or has multiple levels, and there is a need to combine the weights from the
same general design, but with different dichotomizations. Therefore multiple
\code{WeightedDesign}s can be combined if they are identical except for
their \code{@dichotomization} slots.
}
