% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeightedDesign.R
\name{+,WeightedDesign,numeric-method}
\alias{+,WeightedDesign,numeric-method}
\alias{+,numeric,WeightedDesign-method}
\alias{-,WeightedDesign,numeric-method}
\alias{-,numeric,WeightedDesign-method}
\alias{*,WeightedDesign,numeric-method}
\alias{*,numeric,WeightedDesign-method}
\alias{/,WeightedDesign,numeric-method}
\alias{/,numeric,WeightedDesign-method}
\title{\code{WeightedDesign} Ops}
\usage{
\S4method{+}{WeightedDesign,numeric}(e1, e2)

\S4method{+}{numeric,WeightedDesign}(e1, e2)

\S4method{-}{WeightedDesign,numeric}(e1, e2)

\S4method{-}{numeric,WeightedDesign}(e1, e2)

\S4method{*}{WeightedDesign,numeric}(e1, e2)

\S4method{*}{numeric,WeightedDesign}(e1, e2)

\S4method{/}{WeightedDesign,numeric}(e1, e2)

\S4method{/}{numeric,WeightedDesign}(e1, e2)
}
\arguments{
\item{e1}{\code{WeightedDesign} or \code{numeric}}

\item{e2}{\code{numeric} or \code{WeightedDesign}}
}
\description{
\code{WeightedDesign}s do not support addition or subtraction, but do
support all other reasonable operations.
}
