% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeightedStudySpecification.R
\name{subset,WeightedStudySpecification-method}
\alias{subset,WeightedStudySpecification-method}
\alias{[,WeightedStudySpecification-method}
\title{\code{WeightedStudySpecification} subsetting}
\usage{
\S4method{subset}{WeightedStudySpecification}(x, subset)

\S4method{[}{WeightedStudySpecification}(x, i)
}
\arguments{
\item{x}{\code{WeightedStudySpecification} object}

\item{subset}{Logical vector identifying values to keep or drop}

\item{i}{indices specifying elements to extract or replace. See
\code{help("[")} for further details.}
}
\value{
A \code{WeightedStudySpecification} object which is a subsetted
version of \code{x}.
}
\description{
Provides functionality to subset the weights of a
\code{WeightedStudySpecification} object.
}
