% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeightedStudySpecification.R
\name{+,WeightedStudySpecification,numeric-method}
\alias{+,WeightedStudySpecification,numeric-method}
\alias{+,numeric,WeightedStudySpecification-method}
\alias{-,WeightedStudySpecification,numeric-method}
\alias{-,numeric,WeightedStudySpecification-method}
\alias{*,WeightedStudySpecification,numeric-method}
\alias{*,numeric,WeightedStudySpecification-method}
\alias{/,WeightedStudySpecification,numeric-method}
\alias{/,numeric,WeightedStudySpecification-method}
\title{\code{WeightedStudySpecification} Operations}
\usage{
\S4method{+}{WeightedStudySpecification,numeric}(e1, e2)

\S4method{+}{numeric,WeightedStudySpecification}(e1, e2)

\S4method{-}{WeightedStudySpecification,numeric}(e1, e2)

\S4method{-}{numeric,WeightedStudySpecification}(e1, e2)

\S4method{*}{WeightedStudySpecification,numeric}(e1, e2)

\S4method{*}{numeric,WeightedStudySpecification}(e1, e2)

\S4method{/}{WeightedStudySpecification,numeric}(e1, e2)

\S4method{/}{numeric,WeightedStudySpecification}(e1, e2)
}
\arguments{
\item{e1, e2}{\code{WeightedStudySpecification} or \code{numeric} objects}
}
\value{
a \code{WeightedStudySpecification} object
}
\description{
Algebraic operators on \code{WeightedStudySpecification}
objects and numeric vectors. \code{WeightedStudySpecification}s do not
support addition or subtraction.
}
\details{
These are primarily used to either combine weights via
multiplication, or to invert weights. Addition and subtraction are not
supported and will produce errors.
}
