% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeightedDesign.R
\name{weights,WeightedDesign-method}
\alias{weights,WeightedDesign-method}
\title{Extract Weights from WeightedDesign}
\usage{
\S4method{weights}{WeightedDesign}(object, ...)
}
\arguments{
\item{object}{WeightedDesign object}

\item{...}{Ignored}
}
\value{
Weights
}
\description{
Extract Weights from WeightedDesign
}
