% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeightedDesign.R
\name{weights,WeightedDesign-method}
\alias{weights,WeightedDesign-method}
\title{Extract Weights from \code{WeightedDesign}}
\usage{
\S4method{weights}{WeightedDesign}(object, ...)
}
\arguments{
\item{object}{a \code{WeightedDesign} object}

\item{...}{Ignored}
}
\value{
A numeric \code{vector} of the weights
}
\description{
A \code{WeightedDesign} object contains a numeric vector with a
few additional slots, this extracts only the numeric vector.
}
