% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeightedStudySpecification.R
\name{weights,WeightedStudySpecification-method}
\alias{weights,WeightedStudySpecification-method}
\title{Extract Weights from \code{WeightedStudySpecification}}
\usage{
\S4method{weights}{WeightedStudySpecification}(object, ...)
}
\arguments{
\item{object}{a \code{WeightedStudySpecification} object}

\item{...}{Ignored}
}
\value{
A numeric \code{vector} of the weights
}
\description{
A \code{WeightedStudySpecification} object contains a numeric
vector with a few additional slots, this extracts only the numeric vector.
}
